### 2024-05-13 #181 @teald
astrodata
Submitting Author: D.J. Teal (@teald)
All current maintainers: @teald, @chris-simpson, @jehturner
Package Name: astrodata
One-Line Description of Package: Common interface for astronomical data products.
Repository Link: [https://github.com/GeminiDRSoftware/astrodata](https://github.com/GeminiDRSoftware/astrodata)
Version submitted: 2.9.2
Archive: https://zenodo.org/records/14804283
Version accepted: 2.10.1

---

`astrodata` is a package meant to facilitate developing common interfaces for astronomical data formats. Often, specific instruments and models will have different ways of storing their data, including metadata. `astrodata` offers a single interface that uses metadata to resolve these disparate file formats, while enabling common operations and values to share the same interface. The abstraction is meant to be conceptually simple and meaningful to scientists.

Previously, `astrodata` was a core module within the [DRAGONS](https://github.com/GeminiDRSoftware/DRAGONS) package, and has been used for the various instruments that exist at the Gemini Observatory. It has proved useful in consolidating differences in metadata and data formatting between instruments that produce FITS files, which is a common pattern in astronomical data handling. Alongside automating interface selection based on these differences, it also comes with helpful operators and methods out-of-the-box, by extending `astropy`'s `NDData` class.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

    Astronomers & astronomical software developers

    There are no specific packages we are aware of. There is some overlap with the [`gwcs`](https://github.com/spacetelescope/gwcs) package and our `wcs` module, but we are planning to collaborate with that package for future development and to reconsolidate those overlaps. Astrodata offers an interface based on `astropy.nddata.NDData`, but allowing more than one instance to be mapped to the same file (e.g., to multiple sets of FITS extensions).

    N/A
