### 2024-05-14 0034527
Update logo for light mode
