+++ docs/conf.py
@@ -107,14 +107,17 @@ html_theme_options = {
     "body_max_width": "none",
 }
 
+html_logo = "static/logo.svg"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-# html_static_path = ["_static"]
+html_static_path = ["static"]
 
-# TODO: Need to move this file into the directory above it (./_static)
-html_css_files = ["manuals/_static/color_styles.css"]
+html_css_files = [
+    "../manuals/_static/color_styles.css",
+    "css/logo_variables.css",
+]
 
 # Render TODOs; should only be True for the development version.
 todo_include_todos = False
