### 2024-05-14 d0fdd48
Add new logo handling based on github feature
URL describing the feature:
https://github.blog/changelog/2022-05-19-specify-theme-context-for-images-in-markdown-beta/
