+++ README.md
@@ -13,18 +13,15 @@
 `astrodata`
 =============
 
-<!--
-KEEPING HERE FOR REFERENCE
-<img class="logo" align="left" src="docs/static/logo.svg#gh-dark-mode-only" height=200
-style="padding-right: 10; padding-bottom: 10; border: none;">
-
-<img class="logo" align="left" src="docs/static/logo.svg#gh-light-mode-only" height=200
-style="padding-right: 10; padding-bottom: 10; border: none;">
--->
-
 <picture>
   <source media="(prefers-color-scheme: dark)" srcset="docs/static/logo.svg">
-  <img alt="A logo of a stylized blue dragon inside a similarly blue shell. A yellow star lies at the center, together with the dragon shape forming a stylized letter A." src="docs/static/logo_dark.svg">
+  <img
+  alt="A logo of a stylized blue dragon inside a similarly blue shell. A yellow star lies at the center, together with the dragon shape forming a stylized letter A."
+  src="docs/static/logo_dark.svg"
+  align="left"
+  height=200
+  style="padding-right: 10; padding-bottom: 10; border: none;"
+  >
 </picture>
 
 ### Tests
