### 2024-05-14 89ceaf7
Updates to logos for docs, README (#15)
* Adding logos to static folder.

* Update logo for light mode
* Add new logo handling based on github feature

URL describing the feature:
https://github.blog/changelog/2022-05-19-specify-theme-context-for-images-in-markdown-beta/
