### 2024-05-21 675eed1
Update outdated docstring
