### 2024-05-21 ae9913e
Add download_multiple files and tests
