+++ tests/integration/dragons/conftest.py
@@ -1,0 +1,1 @@
+"""Configures the dragons integration tests."""
