### 2024-05-26 #181 @Batalex
Hi there! Thank you for submitting your package for pyOpenSci review. Below are the basic checks that your package needs to pass to begin our review. If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

  **The package installation does not install the dependencies**
  Seems like a BSD3 derivative is ok.

---
---

Nothing to say from the EiC initial perspective, it's a strong submission!
