### 2024-05-28 1e63475
Update cache_purge.yml
Makes the purge manual, which should be good now that there's fewer updates to pages that rely on this cache.
