### 2024-05-28 ace1306
Add weekly tests to testing.yml
