+++ .github/workflows/testing.yml
@@ -7,6 +7,8 @@ on:
     branches:
       - main
   pull_request:
+  schedule:
+      - cron: "20 1 * * sat"
 
 defaults:
   run:
