### 2024-05-29 01a0637
Update CONTRIBUTING.md with flowchart
Also fixes a typo/some wording.
