+++ _sources/index.rst.txt
@@ -1,0 +1,21 @@
+astrodata Documentation
+-----------------------
+
+This is the documentation for astrodata.
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Contents:
+
+   quickstart.rst
+   manuals/index
+   examples/index
+   api_short
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
