+++ _sources/quickstart.rst.txt
@@ -1,0 +1,387 @@
+Quickstart
+----------
+
+This guide is meant to familiarize you with the most important
+concepts of |astrodata|, specifically as someone using |astrodata| to
+work with your own data.
+
+If you are a user getting started with |astrodata|, you will likely
+wnt to start with the |UserManual|. If you want a more detailed overview
+of the library, you can start with the |DeveloperGuide|, which covers the
+majority of topics discussed here in more detail.
+
+Installation
+============
+
+The easiest way to install |astrodata| is to use `pip`:
+
+.. code-block:: bash
+
+    pip install astrodata
+
+This will install the latest stable release of |astrodata|, as well as
+all of its standard dependencies (i.e., not the development dependencies:
+see `Advanced Usage`_).
+
+Example Files
+=============
+
+As |astrodata| is part of the |DRAGONS| software system, there are some testing
+helper functions to download example data for testing purposes. You can
+download files using :func:`~astrodata.testing.download_from_archive`:
+
+.. code-block:: python
+
+    from astrodata.testing import download_from_archive
+
+    # Files we'll use in this example.
+    files = [
+        'N20170614S0201',
+        'N20170614S0202',
+        'N20170614S0203',
+        'N20170614S0204',
+        'N20170614S0205',
+        'N20170613S0180',
+        'N20170613S0181',
+        'N20170613S0182',
+        'N20170613S0183',
+        'N20170613S0184',
+        'N20170615S0534',
+        'N20170615S0535',
+        'N20170615S0536',
+        'N20170615S0537',
+        'N20170615S0538',
+        'N20170702S0178',
+        'N20170702S0179',
+        'N20170702S0180',
+        'N20170702S0181',
+        'N20170702S0182',
+        'bpm_20170306_gmos-n_Ham_22_full_12amp',
+    ]
+    files = [f + '.fits' for f in files]
+
+    # Download the files. We'll be placing these in a directory called 'data'.
+    for f in files:
+        download_from_archive(f, path='data', sub_path='')
+
+These files are from the |DRAGONS| GMOS tutorial, but we'll pretend we don't
+have that and want to work with them using |astrodata| alone. They will be
+stored in the ``./data/`` directory relative to your current working directory.
+
+Opening Files
+=============
+
+The primary way to interact with |astrodata| is through the
+:func:`~astrodata.from_file` function. This function will open a file and
+create an |AstroData| object from it. You can then use this object to access
+the data and metadata in the file.
+
+Let's just open one of the files without any other setup and see what happens:
+
+.. code-block:: python
+
+    import astrodata
+
+    ad = astrodata.from_file('data/N20170614S0201.fits')
+
+    ad.info()
+
+This will print out a summary of the contents of the file. In this case,
+it will look something like this:
+
+.. code-block:: text
+
+    Filename: data/N20170614S0201.fits
+    Tags:
+
+    Pixels Extensions
+    Index  Content                  Type              Dimensions     Format
+    [ 0]   science                  NDAstroData       (2112, 288)    uint16
+    [ 1]   science                  NDAstroData       (2112, 288)    uint16
+    [ 2]   science                  NDAstroData       (2112, 288)    uint16
+    [ 3]   science                  NDAstroData       (2112, 288)    uint16
+    [ 4]   science                  NDAstroData       (2112, 288)    uint16
+    [ 5]   science                  NDAstroData       (2112, 288)    uint16
+    [ 6]   science                  NDAstroData       (2112, 288)    uint16
[…]
