+++ _sources/manuals/index.rst.txt
@@ -1,0 +1,57 @@
+.. Astrodata Master Manual
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+   Manually edited by SC December 2020
+
+################
+Astrodata Manual
+################
+
+.. admonition:: Document ID
+
+   PIPE-USER-120_AstrodataMasterManual
+
+.. toctree::
+   :hidden:
+
+   cheatsheet
+   usermanual/index
+   progmanual/index
+   ../examples/index
+   full_api
+
+This documentation provides different levels of information:
+
+- :doc:`cheatsheet` - A refresher on common astrodata operations
+- :doc:`usermanual/index` - How to code with astrodata
+- :doc:`progmanual/index` - How to code for astrodata
+- :doc:`full_api` - The full API documentation for developers.
+- :doc:`../examples/index` - Examples of how to use astrodata
+
+If you need support related to |astrodata|, see :ref:`ad_support`.
+
+.. raw:: latex
+
+   % Set up the appendix mode and modify the LaTeX toc behavior
+   \appendix
+   \noappendicestocpagenum
+   \addappheadtotoc
+
+
+.. rubric:: Appendix
+
+.. toctree::
+   :maxdepth: 1
+
+   appendix_descriptors
+..   api
+
+.. ******************
+.. Indices and tables
+.. ******************
+
+.. * :ref:`genindex`
+.. * :ref:`modindex`
+.. * :ref:`search`
+
+.. todolist::
