+++ _sources/manuals/usermanual/tags.rst.txt
@@ -1,0 +1,274 @@
+.. tags.rst
+
+.. _tags:
+
+***************
+Astrodata |Tag|
+***************
+
+What is an Astrodata |Tag|?
+===========================
+
+|Tag| is a way to describe the data in an |AstroData| object. Tags are used to
+idenfity the type of |AstroData| object to be created when |open| is called.
+
+A |Tag| is added to an |AstroData| object by defining a function wrapped with
+the :func:`~astrodata.astro_data_tag` decorator.  The function must return a
+:class:`~astrodata.TagSet` object, which describes the behavior of a tag.
+
+For example, the following function defines a tag called "RAW"
+
+.. testsetup::
+
+    from astrodata import TagSet, astro_data_tag
+
+.. code-block:: python
+
+    class RawAstroData(AstroData):
+        @astro_data_tag
+        def _tag_raw(self):
+            """Identify if this is raw data"""
+            if self.phu.get('PROCTYPE') == 'RAW':
+                return TagSet(['RAW'])
+
+Now, if we call |open| on a file that has a PROCTYPE keyword set to "RAW", the
+|AstroData| object will have the "`RAW`" tag
+
+
+.. code-block:: python
+
+    >>> ad = astrodata.open('somefile.fits')
+    >>> ad.tags
+    {'RAW'}
+
+From here, these tag sets can be used to understand what the data is describing
+and how best to process it. It can also contain information about the state of
+processing (e.g., ``RAW`` vs ``PROCESSED``), or any important flags.
+
+.. _ad_tags: :ref:`../progmanual/tags.rst`
+
+These tags are meant to work well with FITS data, using the headers to
+determine what the data is.  However, they can be used with any data type that
+can be described by a set of tags, as long as they are properly defined by the
+developer (see ad_tags_ for more information about developing with |Tag|).
+
+..
+    The Astrodata Tags identify the data represented in the |AstroData| object.
+    When a file on disk is opened with |astrodata|, the headers are inspected to
+    identify which specific |AstroData| class needs to be loaded,
+    :class:`~gemini_instruments.gmos.AstroDataGmos`,
+    :class:`~gemini_instruments.niri.AstroDataNiri`, etc. Based on the class the data is
+    associated with, a list of "tags" will be defined. The tags will tell whether the
+    file is a flatfield or a dark, if it is a raw dataset, or if it has been processed by the
+    recipe system, if it is imaging or spectroscopy. The tags will tell the
+    users and the system what that data is and also give some information about
+    the processing status.
+
+For some examples of tags in production code, see the |gemini_instruments|
+package, which defined a number of |AstroData| derivatives used as part of the
+|DRAGONS| data reduction library for reading as well as processing data.
+
+Using the Astrodata Tags
+========================
+
+**Try it yourself**
+
+Download the data package (:ref:`datapkg`) if you wish to follow along and run the
+examples.  Then ::
+
+    $ cd <path>/ad_usermanual/playground
+    $ python
+
+Before doing anything, you need to import |astrodata| and the Gemini instrument
+configuration package (|gemini_instruments|).
+
+::
+
+    >>> import astrodata
+    >>> import gemini_instruments
+
+Let us open a Gemini dataset and see what tags we get::
+
+    >>> ad = astrodata.open('../playdata/N20170609S0154.fits')
+    >>> ad.tags
[…]
