+++ _sources/examples/generic_examples/index.rst.txt
@@ -1,0 +1,30 @@
+.. _generic_examples:
+
+Generic Examples
+================
+
+This section contains examples using |astrodata| with no specific instrument
+or pipeline in mind. The examples are intended to demonstrate the basic
+functionality of |astrodata| and to provide a starting point for users
+who are new to the package.
+
+Many of these patterns will apply broadly to anything derived from
+|AstroData|. However, no "real" data is used in these examples, and
+they are not intended to represent how any given reduction "should"
+be done to achieve science quality data.
+
+.. _generic_examples_user:
+
+User Examples
+-------------
+
+.. toctree::
+    :maxdepth: 2
+    :glob:
+
+    usermanual_structure_examples/1_example_structure.rst
+
+.. _generic_examples_developer:
+
+Developer Examples
+------------------

[…]
