+++ .github/workflows/nox_testing.yml
@@ -1,0 +1,52 @@
+name: "Nox Tests"
+
+on:
+    push:
+        branches:
+        - main
+    pull_request:
+
+defaults:
+    run:
+        shell: bash
+
+jobs:
+    tests:
+        name: "Python ${{ matrix.python-version }} on ${{ matrix.os }}"
+        runs-on: "${{ matrix.os }}"
+
+        strategy:
+            fail-fast: false
+            matrix:
+                os:
+                    - ubuntu-latest
+                    - macos-latest
+                    - windows-latest
+                python-version:
+                    - "3.10"
+                    - "3.11"
+                    - "3.12"
+
+        steps:
+            - name: "Check out the repo"
+              uses: "actions/checkout@v4"
+
+            - name: "Set up Python"
+              uses: "actions/setup-python@v5"
+              with:
+                  python-version: "${{ matrix.python-version }}"
+
+            - name: "Install dependencies"
+              run: |
+                  python -m pip install --upgrade pip
+                  python -m pip install nox
+
+            - name: "Run nox for ${{ matrix.python-version }}"
+              run: |
+                  python -m nox -r
+
+            - name: "Upload coverage data"
+              uses: actions/upload-artifact@v3
+              with:
+                  name: covdata
+                  path: .coverage.*
