+++ pyproject.toml
@@ -45,6 +45,7 @@ pytest-cov = "^4.1.0"
 pytest-doctestplus = "^1.2.1"
 tox = "^4.14.1"
 hypothesis = "^6.99.13"
+nox = "^2024.4.15"
 
 [build-system]
 requires = ["poetry-core"]
