+++ tests/integration/dragons/test_niri_imaging.py
@@ -1,0 +1,249 @@
+"""Tests for the NIRI imaging mode.
+
+.. _DRAGONS_NIRIIMG_TUTORIAL: https://dragons.readthedocs.io/\
+projects/niriimg-drtutorial/en/v3.2.0/ex1_niriim_extended_dataset.html
+"""
+
+import glob
+import importlib
+import os
+
+import pytest
+
+
+from astrodata.testing import expand_file_range, download_multiple_files
+
+
+@pytest.fixture(scope="session")
+def niri_imaging_data_star_field_files():
+    """Retrieve NIRI imaging data for the star field tutorial."""
+    files = [
+        "N20160102S0270-274",
+        "N20160102S0275-279",
+        "N20160102S0423-432",
+        "N20160102S0373-382",
+        "N20160102S0363-372",
+        "N20160103S0463-472",
+        "N20160102S0295-299",
+        "bpm_20010317_niri_niri_11_full_1amp.fits",
+    ]
+
+    expanded_files = []
+
+    for file in files:
+        if not file.startswith("bpm"):
+            expanded_files += expand_file_range(file)
+            continue
+
+        expanded_files += [file]
+
+    return expanded_files
+
+
+@pytest.fixture(scope="session")
+def _downloaded_niri_imaging_data_star_field(
+    tmpdir_factory, niri_imaging_data_star_field_files
+):
+    """Download NIRI imaging data for the star field tutorial."""
+    tmpdir = tmpdir_factory.mktemp("niri_imaging_data_star_field")
+
+    data = download_multiple_files(
+        niri_imaging_data_star_field_files,
+        path=tmpdir,
+        sub_path="",
+    )
+
+    return data
+
+
+@pytest.fixture
+def niri_imaging_data_star_field(
+    tmp_path, _downloaded_niri_imaging_data_star_field
+):
+    """This copies files from another fixture
+    (_downloaded_niri_imaging_data_star_field), and copies those into a path
+    for the current test.
+    """
+    # Copy the files from the temporary directory to the test directory instead
+    # of re-downlaoding them.
+    data = {}
+
+    for file, path in _downloaded_niri_imaging_data_star_field.items():
+        data[file] = tmp_path / file
+
+        with open(data[file], "wb") as f:
+            with open(path, "rb") as f2:
+                f.write(f2.read())
+
+    return data
+
+
+@pytest.mark.dragons
+def test_correct_astrodata():
+    """Test if the astrodata package is being tested."""
+    # Importing astrodata as if used with DRAGONS, instead of the top
+    # import, for clarity/to test it explicitly.
+    astrodata = importlib.import_module("astrodata")  # noqa: F811
+    assert astrodata.from_file
+
+
+@pytest.mark.filterwarnings("ignore:use 'astrodata.from_file'")
+@pytest.mark.filterwarnings("ignore:Renamed to 'as_iraf_section'")
+@pytest.mark.filterwarnings("ignore:Renamed to add_class")
+@pytest.mark.filterwarnings("ignore:Renamed to 'windowed_operation'")
+@pytest.mark.filterwarnings("ignore")
+@pytest.mark.dragons
+def test_niri_imaging_tutorial_star_field(
+    use_temporary_working_directory,
+    niri_imaging_data_star_field,
+):
+    """Test based on the DRAGONS NIRI imaging tutorial.
+
+    This does **not** follow the tutorial directly. It is just testing based on
+    the tutorial, and should be updated if the tutorial changes.
+
+    Notably, importing is a bit different, and the tutorial uses a different
+    method to get the data.
+
+    Link: `DRAGONS_NIRIIMG_TUTORIAL`_
+    """
[…]
