### 2024-06-13 cb99b9d
Make stdout optional for downloads
