+++ .github/workflows/testing.yml
@@ -14,7 +14,7 @@ defaults:
 
 jobs:
     tests:
-        name: "Python ${{ matrix.python-version }} on ${{ matrix.os }}"
+        name: "Testing on ${{ matrix.os }}"
         runs-on: "${{ matrix.os }}"
 
         strategy:
@@ -22,12 +22,9 @@ jobs:
             matrix:
                 os:
                     - ubuntu-latest
-                    - macos-latest
-                    - windows-latest
-                python-version:
-                    - "3.10"
-                    - "3.11"
-                    - "3.12"
+                    # TODO:
+                    # - macos-latest
+                    # - windows-latest
 
         steps:
             - name: "Check out the repo"
@@ -36,12 +33,12 @@ jobs:
             - name: "Set up Python"
               uses: "actions/setup-python@v5"
               with:
-                  python-version: "${{ matrix.python-version }}"
+                  python-version: "3.11"
 
             - name: "Install dependencies"
               run: |
                   python -m pip install --upgrade pip
-                  python -m pip install nox
+                  python -m pip install nox poetry
 
             - name: "Get public IP address of runner"
         # Need to handle this for Windows, MacOS, and linux
@@ -55,7 +52,7 @@ jobs:
               echo "Linux runner IP: $(curl -s https://ipinfo.io/ip)"
             fi
 
-      - name: "Run nox for ${{ matrix.python-version }}"
+      - name: "Run nox"
               run: |
                   python -m nox -r
 
