### 2024-06-25 49056e4
Add to niri tests, update test config
