### 2024-07-15 46d0ae3
Fix typo in README
licence -> license
