### 2024-07-15 GeminiDRSoftware/astrodata#16 @teald
[BUG] Intermittent failure of download in testing workflow
**Describe the bug**
During some test workflows (e.g., [this workflow run](https://github.com/GeminiDRSoftware/astrodata/actions/runs/9916487175)), `astrodata.testing.download_from_archive` will get a timeout error while attempting to read in a file.

**To Reproduce**
Unknown, it's unclear if this is a problem with `astrodata`, GitHub Actions/the workflow file, or something external.

**Expected behavior**
Downloads should complete without timing out.

**Example code**
See e.g., [workflow run](https://github.com/GeminiDRSoftware/astrodata/actions/runs/9916487175)

 - OS: variable (occurs on ubuntu, macos, windows intermittently)

**Additional context**
It seems like [recent workflow runs](https://github.com/GeminiDRSoftware/astrodata/actions/runs/9945000816) may be working again, but it's unclear why.
