### 2024-07-15 28be095
Use Path objects in noxfile
