### 2024-07-15 dbe8f92
Linting noxfile
