### 2024-07-15 cfc3f18
Update unit test selection
It's more obvious to use the directory tree for separating classes of
tests than using marks (now that there are also integration tests).
