### 2024-07-15 ddb37ed
Explicitly ignore return results for session runs.
