### 2024-07-15 #181 @aaryapatil
Need to further check the above two points.

    - url isn't present.

        - please include
        - please include

   - For a first time reader going through the README, it might be worthwhile to expand upon the utility of `astrodata` beyond using `astropy.io.fits` alone. The current text only briefly mentions the extendability of `astrodata`.

CITATION.md is present but not referenced in the README.

---

There are a few typos present in the documentation:

- In Quickstart:
    - you will likely ~wnt~ want to start with the [User Guide](https://geminidrsoftware.github.io/astrodata/manuals/usermanual/index.html)
    -  there are some ~testing~ helper functions to download example data for testing purposes
    - ~from~ with development dependencies
