### 2024-07-15 a1ab7b6
Adjust for updates to test directory structure
