### 2024-07-16 b0699bb
Add ruff config to pyproject.toml
Start of the process of migrating these settings from
pre-commit/workflows into the pyproject.toml file.
