+++ pyproject.toml
@@ -106,3 +106,13 @@ exclude_lines = [
   # Don't worry about breakpoints (from testing file)
   "breakpoint",
 ]
+
+
+# TODO: Expand ruff settings
+[tool.ruff.lint]
+ignore = [
+  "RET504",
+]
+
+[tool.ruff.lint.pydocstyle]
+convention = "numpy"
