### 2024-07-16 GeminiDRSoftware/astrodata#18 @mwcraig
[DOC] Quickstart text describes `airmass` descriptor as "central wavelength"
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

At a couple places in the quickstart you mention adding a descriptor for "central wavelength" but the descriptor added in the quickstart is for "airmass". See [here](https://github.com/GeminiDRSoftware/astrodata/blob/main/docs/quickstart.rst?plain=1#L292) and [here](https://github.com/GeminiDRSoftware/astrodata/blob/main/docs/quickstart.rst?plain=1#L314)

## Solution

I assume the intent is for the text to change, not the code....

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
