### 2024-07-16 GeminiDRSoftware/astrodata#18 @mwcraig
Also, hello 👋, I'm one of the pyOpenSci/astropy affiliated package reviewers. Excited about learning more about this package!
