### 2024-07-16 GeminiDRSoftware/astrodata#19 @mwcraig
[DOC] Quickstart suggestion -- add an example of tags
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

Based on a brief read of the rest of the documentation it looks like tags are an important concept/tool in `astrodata`. If it is straightforward to do, adding an example of tags in the quickstart would be great.

If it isn't feasible, then you could remove [this import](https://github.com/GeminiDRSoftware/astrodata/blob/main/docs/quickstart.rst?plain=1#L179).

## Solution
[If applicable, your solution to the issue]

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
