### 2024-07-16 338586d
Fix typos in quickstart.rst
