### 2024-07-16 068853c
Remove argument
In a recent version of poetry (1.8.3) no positional arguments are allowed in install
