### 2024-07-16 GeminiDRSoftware/astrodata#22 @mwcraig
[DOC] Add note to "Installing development dependencies" section of README about editable installs
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

I am a frequent user of pip but an very infrequent  (n=1) user of poetry. I recommend adding a short note to `README.md` that indicates that the install is editable, i.e. roughly equivalent to the `--editable` argument to pip.

## Solution
[If applicable, your solution to the issue]

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [x] This problem is minor/does not require immediate fixing.
