### 2024-07-16 GeminiDRSoftware/astrodata#23 @mwcraig
[DOC] There are three different versions of how to do a development install
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

Instructions for doing a dev install show up in three places:

- README.md, where the command is `poetry install`
- quickstart.rst, where the command is `poetry install .` (fixed in #21 )
- manuals/usermanual/intro.rst, where the command is `pip install -e .`

## Solution

It seems like the dev install should be done with one tool -- either `poetry` or `pip`.

### Urgency
- [x] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
