### 2024-07-16 0fa8411
Update README.md poetry description.
