### 2024-07-16 GeminiDRSoftware/astrodata#25 @mwcraig
[DOC] Remove dev install instructions from quickstart, add separate install doc at top level
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

Right now developer instructions show up in several places (see #23).

I'd recommend making an installation-specific document that shows up in the documentation sidebar at the same level as the quickstart. That installation-specific document should have a section for a user install and a section for a developer install, with a note like that in #22 for folks unfamiliar with `poetry`.

I think this would have a couple of advantages:

- It would make it easier to find both flavors of installation in the documentation.
- It would let you remove any installation instructions from the quickstart (especially the dev instructions, which I don't think belong in a quickstart) and the user manual introduction.

## Solution
[If applicable, your solution to the issue]

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
