### 2024-07-16 GeminiDRSoftware/astrodata#26 @mwcraig
[DOC] Progress bar/indicator for file download in quickstart would be nice
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

For reasons I don't understand the download of the files for the quickstart is taking roughly forever on my machine (this is a me problem, not an astrodata problem).

It would be nice to have some indication of how far the downloader has gotten. DSoesn't need to be a fancy tqdm-style progress bar, a simple `print(f"On file {file}")` would be helpful.

## Solution
[If applicable, your solution to the issue]

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [x] This problem is minor/does not require immediate fixing.
