### 2024-07-16 GeminiDRSoftware/astrodata#26 @teald
We're actually experiencing some issues with downloads right now (related to Issue #16). It doesn't seem to be `astrodata`, as it's happening for other software our team uses. So it may not be an issue on your end either!

That said, I agree a progress bar (or an obvious option specifying one) would be nice in general, thanks for the suggestion!There's a progress bar generated/that can be generated by the library we use, so I'll look into making that something that can be more easily made visible.
