### 2024-07-16 GeminiDRSoftware/astrodata#27 @mwcraig
[DOC] Links to `ad_tags` in user manual are broken
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

There are a couple of references to `ad_tags` in the [tags section of the user manual](https://geminidrsoftware.github.io/astrodata/manuals/usermanual/tags.html).

Clicking on the `ad_tags` link gives me a 404/file not found error.

## Solution
[If applicable, your solution to the issue]

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
