### 2024-07-16 GeminiDRSoftware/astrodata#28 @mwcraig
[DOC] The "quick example" in the user manual intro does not work
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

The [quick example](https://geminidrsoftware.github.io/astrodata/manuals/usermanual/intro.html#a-quick-example) in the user's manual introduction has users import `astrodata.testing` but that is not installed with a standard `pip install astrodata`, so the import fails.

## Solution

Either explicitly require an installation that includes `astrodata.testing` or change the quick example.

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
