### 2024-07-16 GeminiDRSoftware/astrodata#29 @mwcraig
[DOC] Accessibility issues in dark mode
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description

There are a few issues with the documentation when viewed in dark mode.

1. Class names cannot be read because they are in black text (pic 1 below)
2. The python prompts very hard to see in dark mode (see pic 2 below)
3. Whether class and function names can be read depends a lot on screen brightness (see pic 2 below). On my laptop screen on battery power while riding in a vehicle during the daytime they were really hard to read. On the monitor of my home computer or if I turned up the brightness of the laptop screen I could see them.

## Solution

One short term fix would be to disable dark mode if possible. An alternative is to check whether there is additional css for dark mode that should be applied to your code snippets.

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
