### 2024-07-16 GeminiDRSoftware/astrodata#29 @teald
There is an issue about this with the dark theme we use, but it's not been worked on yet as far as I can tell.
+ https://github.com/MrDogeBro/sphinx_rtd_dark_mode/issues/27

It is probably best to add some CSS or find another theme, as suggested. Thanks for catching this (especially the class names in the API reference)! I'm surprised I hadn't yet, given I mostly use dark mode 😅
