+++ docs/examples/generic_examples/usermanual_structure_examples/1_example_structure.rst
@@ -1,0 +1,4 @@
+.. _user_structure_examples:
+
+Global and Extension-Specific Metadata
+======================================
