### 2024-07-16 GeminiDRSoftware/astrodata#30 @teald
[FEATURE] Workflow to enforce linting/styling
**Is your feature request related to a problem? Please describe or write "N/A".**
N/A

**Describe the solution you'd like**

Right now the linting is done as part of the `pre-commit` process, controlled in [`pre_commit_config.yaml`](https://github.com/GeminiDRSoftware/astrodata/blob/46d0ae33a5a9a8d227c2f415dd4c882117537096/.pre-commit-config.yaml). This will not run in certain scenarios:

* Using github to modify files.
* Forgetting to run `pre-commit install` before committing to the repo.

As part of the PyOpenSci review process, this was noted by one of the reviewers (see **Additional Context** for a link to the submission).

The most straightforward solution is to have a workflow run `pre-commit` on target files (similar to how pre-commit itself works normally).

N/A

**Additional context**
* https://github.com/pyOpenSci/software-submission/issues/181
