### 2024-07-17 GeminiDRSoftware/astrodata#16 @teald
This (external) issue has been resolved.
