### 2024-07-17 63ae214
Update wording of poetry instructions
These used to be comments, but the phrasing doesn't flow as well for
paragraphs.
