### 2024-07-17 f04ae46
Add developer installation and testing instructions.
