+++ docs/index.rst
@@ -10,6 +10,7 @@ This is the documentation for astrodata.
    quickstart.rst
    manuals/index
    examples/index
+   developer/index
    api_short
 
 
