### 2024-07-17 59898b9
Remove outdated developer installation instructions
