### 2024-07-17 464ecc3
Add reference to developer install guide
