### 2024-07-17 d6c2c1e
Add reference to developer install guide
