### 2024-07-17 963865e
Add clarification to programming manual introduction.
