### 2024-07-18 fb4ff38
Add copy/paste, additional information to dev docs
Also reorganized the headers in the doc since they were confusing on the
sidebar. If this document gets any longer it needs to be split apart.
