### 2024-07-18 1b4f6d8
Update README.md with citation information
