### 2024-07-18 b057511
Update README.md with citation information
