### 2024-07-18 3e79a88
Change central wavelength to airmass.
