### 2024-07-18 f13b5b5
Change variable names to be more clear.
