### 2024-07-18 bf9aea4
Update quickstart to use less common path for data
