### 2024-07-19 56bc918
Update README.md
+ Add badges requested by pyopensci reviewers
+ Small wording change
+ Move pypi version from misleading "Build" section to the topmost line (for ease of verifying the latest version).
