### 2024-07-20 GeminiDRSoftware/astrodata#33 @aaryapatil
[DOC] Fix docs build warnings
Thank you for submitting an issue related to our documentation! *If this is a minor change or something you are willing to modify directly yourself, your help would be greatly appreciated!* Please modify the file using your preferred method (e.g., [directly in github](https://docs.github.com/en/repositories/working-with-files/managing-files/editing-files) or by [making a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request)). Otherwise, continue on with this form.

## Description
The following three docs build warnings are due to references to example documentation that is currently unavailable. Additionally, there is an `Example Location` subsection in `The AstroData Object` [manual](https://geminidrsoftware.github.io/astrodata/manuals/usermanual/structure.html#the-astrodata-object), which has a broken reference to `user_structure_examples`, which I believe should be removed. It breaks the flow of the text in the manual section; if you want to provide the examples location, you can just add it at the end or in text to the previous subsection.

> /Users/patil/Desktop/astrodata/docs/examples/generic_examples/index.rst:21: WARNING: toctree contains reference to nonexisting document 'examples/generic_examples/usermanual_structure_examples/1_example_structure'
> /Users/patil/Desktop/astrodata/docs/examples/index.rst:11: WARNING: toctree contains reference to nonexisting document 'examples/quickstart/index'
> /Users/patil/Desktop/astrodata/docs/manuals/usermanual/structure.rst:22: WARNING: undefined label: 'user_structure_examples'

## Solution
Populating the examples.

### Urgency
- [ ] This problem should be fixed quickly (urgent)
- [ ] This problem is minor/does not require immediate fixing.
