### 2024-07-22 ab47524
Update README.md to use python -m pip
