### 2024-07-23 GeminiDRSoftware/astrodata#16 @teald
The macos runners are having problems related to this issue.
