### 2024-07-24 fc69147
ci: :loud_sound: Capture the runner IP
