### 2024-07-24 e53661a
ci: 🐛 Remove flag from stalling windows command.
