### 2024-07-24 GeminiDRSoftware/astrodata#16 @teald
This has been resolved. Also, IP addresses are now reported in the testing runs (see PR #34)
