### 2024-07-25 daab3f2
test(dragons): 🧪 Add NIRI and GMOS tests, now working.
