### 2024-07-25 afebe0f
fix: :construction_worker: Fix post-rebase syntax issue in testing.yml
