### 2024-07-25 73a98f1
ci: :green_heart: Allow for manual runs.
