+++ .github/workflows/testing.yml
@@ -7,6 +7,7 @@ on:
     pull_request:
     schedule:
         - cron: "20 1 * * sat"
+    workflow_dispatch:
 
 defaults:
     run:
