+++ .github/workflows/testing.yml
@@ -34,7 +34,10 @@ jobs:
             - name: "Set up Python"
               uses: "actions/setup-python@v5"
               with:
-                  python-version: "3.11"
+                  python-version: |
+                    3.10
+                    3.11
+                    3.12
 
             - name: "Install dependencies"
               run: |
