### 2024-07-25 0ff67b5
ci(dragons): 🧪 add DRAGONS tests to workflow.
