### 2024-07-26 99a681a
test(dragons): 🧪 Add DRAGONS dev version tests
This adds a new testing session in nox that pulls the main DRAGONS
branch and tests on it explicitly, rather than only testing on the
release. This will be important for migrating changes to astrodata over
from DRAGONS.

Currently fails to run.
