+++ .github/workflows/testing.yml
@@ -33,6 +33,11 @@ jobs:
                 auto-activate-base: false
                 python-version: "3.12"
 
+            - name: Set up Homebrew (linux/mac only)
+              if: startsWith(matrix.os, 'ubuntu') || startsWith(matrix.os, 'macos')
+              id: set-up-homebrew
+              uses: Homebrew/actions/setup-homebrew@master
+
             - name: "Check out the repo"
               uses: "actions/checkout@v4"
 
@@ -69,6 +74,7 @@ jobs:
             - name: "DRAGONS integration tests"
               run: |
                 if [[ "$RUNNER_OS" != "Windows" ]]; then
+                  brew install sextractor
                   python -m nox -r -s dragons_calibration
                   python -m nox -r -s dragons_release_tests
                 fi
