### 2024-07-26 ef05b02
ci(dragons): 🛠️ Remove redundant calibration test call.
