+++ .github/workflows/testing.yml
@@ -75,7 +75,6 @@ jobs:
               run: |
                 if [[ "$RUNNER_OS" != "Windows" ]]; then
                   brew install sextractor
-                  python -m nox -r -s dragons_calibration
                   python -m nox -r -s dragons_release_tests
                 fi
 
