+++ .github/workflows/testing.yml
@@ -76,6 +76,7 @@ jobs:
                 if [[ "$RUNNER_OS" != "Windows" ]]; then
                   brew install sextractor
                   python -m nox -r -s dragons_release_tests
+                  python -m nox -r -s dragons_dev_tests
                 fi
 
             - name: "Upload coverage data"
