### 2024-07-29 c35e131
test(dragons): 🐛 Update development tests to use conda for sextractor.
