+++ .github/workflows/testing.yml
@@ -27,11 +27,15 @@ jobs:
                     - windows-latest
 
         steps:
-            - uses: conda-incubator/setup-miniconda@v3
-              with:
-                auto-update-conda: true
-                auto-activate-base: false
-                python-version: "3.12"
+            - name:  Set up conda (linux/mac only)
+              if: startsWith(matrix.os, 'ubuntu') || startsWith(matrix.os, 'macos')
+              id: set-up-conda
+              run:
+                  if [[ "$RUNNER_OS" == "Linux" ]]; then
+                    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+                  elif [[ "$RUNNER_OS" == "macOS" ]]; then
+                    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh
+                  fi
 
             - name: Set up Homebrew (linux/mac only)
               if: startsWith(matrix.os, 'ubuntu') || startsWith(matrix.os, 'macos')
