### 2024-07-29 0a849ec
ci(dragons): 🐛 Fix syntax issues in yaml file, use curl for macos.
