+++ .github/workflows/testing.yml
@@ -23,24 +23,24 @@ jobs:
             matrix:
                 os:
                     - ubuntu-latest
-                    - macos-latest
-                    - windows-latest
+                    #- macos-latest
+                    #- windows-latest
 
         steps:
             - name:  Set up conda (linux/mac only)
               if: startsWith(matrix.os, 'ubuntu') || startsWith(matrix.os, 'macos')
-              id: set-up-conda
-              run:
+              run: |
                   if [[ "$RUNNER_OS" == "Linux" ]]; then
-                    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+                    mkdir -p ~/miniconda3
+                    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
+                    bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
+                    rm -rf ~/miniconda3/miniconda.sh
                   elif [[ "$RUNNER_OS" == "macOS" ]]; then
-                    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh
+                    mkdir -p ~/miniconda3
+                    curl https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-arm64.sh -o ~/miniconda3/miniconda.sh
+                    bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
+                    rm -rf ~/miniconda3/miniconda.sh
                   fi
-
-            - name: Set up Homebrew (linux/mac only)
-              if: startsWith(matrix.os, 'ubuntu') || startsWith(matrix.os, 'macos')
-              id: set-up-homebrew
-              uses: Homebrew/actions/setup-homebrew@master
 
             - name: "Check out the repo"
               uses: "actions/checkout@v4"
