+++ .github/workflows/testing.yml
@@ -23,8 +23,8 @@ jobs:
             matrix:
                 os:
                     - ubuntu-latest
-                    #- macos-latest
-                    #- windows-latest
+                    - macos-latest
+                    - windows-latest
 
         steps:
             - name:  Set up conda (linux/mac only)
