### 2024-07-29 d1d2857
fix(dragons): 🐛 Initialize conda shell.
