+++ .github/workflows/testing.yml
@@ -35,11 +35,13 @@ jobs:
                     wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
                     bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
                     rm -rf ~/miniconda3/miniconda.sh
+                    ~/miniconda3/bin/conda init --all
                   elif [[ "$RUNNER_OS" == "macOS" ]]; then
                     mkdir -p ~/miniconda3
                     curl https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-arm64.sh -o ~/miniconda3/miniconda.sh
                     bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
                     rm -rf ~/miniconda3/miniconda.sh
+                    ~/miniconda3/bin/conda init --all
                   fi
 
             - name: "Check out the repo"
