### 2024-07-29 28c2aa0
ci(dragons): 🧹 Tidy dragons dev install sequence
