### 2024-07-30 faeec5a
ci: :bug: Fix issue with git+ installs
This was caused by a conflict when installing both python and conda in the same action (probably some library thing).
