### 2024-07-30 482f15d
ci: :bug: Use more recent version of `setup-miniconda` (v2->v3)
