+++ .github/workflows/integration_tests.yml
@@ -27,7 +27,7 @@ jobs:
                     - windows-latest
 
         steps:
-            - uses: conda-incubator/setup-miniconda@v2
+            - uses: conda-incubator/setup-miniconda@v3
 
             - name: "Check out the repo"
               uses: "actions/checkout@v4"
