### 2024-07-30 ea2eccc
ci: :bug: Remove windows from DRAGONS integration test runs
