### 2024-07-30 e52de05
ci: :art: Rename testing.yml workflow
This may break badges when it is published.
