+++ .github/workflows/testing.yml
@@ -1,4 +1,4 @@
-name: "Nox Tests"
+name: "Unit Tests"
 
 on:
     push:
