### 2024-07-30 eccf9e1
ci: :art: Move dev tests to their own step
