### 2024-07-30 a545c35
test: :art: Consolidate conda installs
