### 2024-07-30 9b42aae
test: :loud_sound: Increase test verbosity
