### 2024-07-30 65d9155
test: :loud_sound: Autouse proper log handling for DRAGONS
Otherwise, stdout gets double messaging (known issue in DRAGONS).
