### 2024-07-30 8debe1a
ci: :green_heart: Catch bad OS Names for DRAGONS runs
