+++ .github/workflows/integration_tests.yml
@@ -56,6 +56,9 @@ jobs:
                     CONDA_SUBDIR="linux-64" conda create -n "test_env" python=3.10 -y
                   elif [[ "$RUNNER_OS" == "macOS" ]]; then
                     CONDA_SUBDIR="osx-64" conda create -n "test_env" python=3.10 -y
+                  else
+                    echo "Unsupported OS: $RUNNER_OS"
+                    exit 1
                   fi
 
                   conda activate test_env
