### 2024-07-30 9562145
ci: :white_check_mark: Add back in OS's for unit testing
