+++ .github/workflows/testing.yml
@@ -23,8 +23,8 @@ jobs:
             matrix:
                 os:
                     - ubuntu-latest
-                    # TODO: - macos-latest
-                    # TODO: - windows-latest
+                    - macos-latest
+                    - windows-latest
 
         steps:
             - name: "Check out the repo"
