### 2024-07-30 834e314
ci: :wrench: Update the global conda config in the `osx` runners.
