+++ .github/workflows/integration_tests.yml
@@ -55,7 +55,9 @@ jobs:
                   if [[ "$RUNNER_OS" == "Linux" ]]; then
                     CONDA_SUBDIR="linux-64" conda create -n "test_env" python=3.10 -y
                   elif [[ "$RUNNER_OS" == "macOS" ]]; then
-                    CONDA_SUBDIR="osx-64" conda create -n "test_env" python=3.10 -y
+                    # Set the conda subdirectory to osx-64
+                    conda config --set subdir osx-64
+                    conda create -n "test_env" python=3.10 -y
                   else
                     echo "Unsupported OS: $RUNNER_OS"
                     exit 1
