### 2024-07-30 27dc8b6
fix: :bug: Fix issue with config command/subdir permanency
