### 2024-07-30 60cf400
test: :white_check_mark: Add configuration step for macos testing.
This is needed because the env settings to not always propogate down to the conda environemnts generated by `nox`.
