### 2024-07-31 70dc3dd
ci: :sparkles: Add linting to pushes, using pre-commit.
This will execute on every push. Assuming pre-commit has been configured correctly for the dev environment, this should have already been run when the commit was made locally for only the modified files.
