### 2024-07-31 3b3388b
🐛 Ignore no-commit-to-branch for linting action
