### 2024-07-31 26ede3d
test: :technologist: Remove the `dragons/astrodata` dir directly prior to build
This is a quick implementation for the future---this will ensure that any relative imports are still caught when testing astrodata on the development version of DRAGONS. It does *not* solve this problem for the release version, but that's less of an issue for now.
