### 2024-07-31 1ba369c
ci: :construction: Verbose output for nox during github actions.
