### 2024-07-31 3ff7a21
Revert "ci: :construction: Try updating session venv params passed to conda."
This reverts commit 434bcf73f63d169f925b656f2fc3f3b8a2a1fdc2.
