### 2024-07-31 f0ebfb2
ci: :construction: Rearrange subdir invocation.
