### 2024-07-31 7d07b62
ci: :construction: Remove configuration step entirely.
