### 2024-07-31 d2549a7
ci(dragons): :construction: Add poetry install to workflow
