### 2024-07-31 ed6f014
ci(dragons): :construction_worker: Remove macos worker request
This is now covered in Issue #39.
