### 2024-08-01 c5d29ee
Testing upgrade (#35)
Squashing 68 commits that update the testing setup to use nox. See PR #35 for more information.

The commit messages for the squashed commits are listed below (in case they're useful):

* Update outdated docstring

* Add file range expansion to testing module

* Add download_multiple files and tests

* Initial integration testing commit

Includes updates to astrodata/testing.py as well which enable multiple
files to be downloaded asynchronously using threads or processes.

Ideally, that would eventually just be async, but invoking an executor
is easier and gets the job do
[…]
