### 2024-08-01 6d99a6a
docs: :white_check_mark: Remove outdated doctest configuration file reference.
This file is no longer used for configuration, and was causing an error.
