### 2024-08-01 d2faa99
ci(docs): ♻️ adjust documentation build times in workflow
