### 2024-08-01 9bd6d80
feat(dev): :sparkles: Add docs session to nox
It's now possible to build the documentation using only `nox -s docs`.
