### 2024-08-01 71533cd
docs: :bug: Install astrodata when building docs
