### 2024-08-01 490febe
docs: :loud_sound: Add nitpicky flag to docs build
