### 2024-08-01 7110ebe
docs: 🛠️ Fix central wavelength, data path in quickstart
