### 2024-08-01 98df823
test(dragons): ♻️ downgrade numpy for DRAGONS tests.
