+++ _sources/manuals/progmanual/intro.rst.txt
@@ -1,0 +1,213 @@
+.. intro.rst
+
+.. _intro_progmanual:
+
+************
+Introduction
+************
+
+
+.. note::
+    This guide discusses programming your own data interfaces using
+    |astrodata|. If you instead plan to develop the |astrodata| package
+    itself, please see |DeveloperInstall| before starting this guide to install
+    developer dependencies.
+
+AstroData is a Python package that provides a common interface to astronomical
+data. Originally part of the |DRAGONS| package, the data reduction package
+developed at the Gemini Observatory, it has been split into its own package to
+allow its use in other projects and to be developed by a wider, public core to
+suit the needs of data reduction across the field of astronomy.
+
+AstroData's common interface, the |AstroData| class, is used to abstract the
+details of any given data into a set of |Tags|, which can be used to resolve
+the properties and reduction requirements of a given data file (most commonly,
+a multi-extension FITS file). The |AstroData| class is also used to provide
+methods to access the data in a consistent manner, regardless of the
+underlying data format.
+
+The |AstroData| class is not intended to be used directly by the user. Instead,
+the user should use the :func:`~astrodata.from_file` function, which will return an
+|AstroData| object. The :func:`~astrodata.from_file` function will determine the type of
+data file being opened and return the appropriate subclass of |AstroData|.
+
+For the programmer using |AstroData| to develop a data reduction pipeline, the
+|AstroData| class should be subclassed to provide the functionality required
+and to register the new class with the :func:`~astrodata.from_file` function.
+
+Several examples may be found throughout the documentation (see |Examples|). A
+simple example is shown below as a complete, executable introduction.
+
+.. code-block:: python
+
+    >>> # Defining an AstroData subclass
+    >>> from astrodata import (
+    ...     AstroData,
+    ...     astro_data_tag,
+    ...     astro_data_descriptor,
+    ...     TagSet
+    ... )
+    >>> import astrodata
+
+    >>> class MyAstroData(AstroData):
+    ...     @staticmethod
+    ...     def _matches_data(source):
+    ...         # This method is used by astrodata.from_file to determine if this
+    ...         # class should be used to open the data. It should return True
+    ...         # if the data is of the correct type, False otherwise.
+    ...
+    ...         # E.g., if file limited by FITS standard or not for instrument
+    ...         # keyword.
+    ...         instrument_tags = {'INSTRUME', 'INSTRUMENT'}
+    ...
+    ...         for tag in instrument_tags:
+    ...             if tag in source[0].header:
+    ...                 return source[0].header.get(tag).upper() == 'MY_INSTRUMENT'
+    ...
+    ...         # Could return None by default since it's Falsey, but this is more
+    ...         # explicit and follows typing expectations.
+    ...         return False
+    ...
+    ...     @astro_data_tag
+    ...     def my_tag(self):
+    ...         # This method is used to define a new tag. It should return
+    ...         # a string that will be used as the tag name. The method name
+    ...         # is used as the tag name by default, but this can be overridden
+    ...         # by passing a name to the decorator, e.g.:
+    ...         # @astro_data_tag(name='my_tag_name')
+    ...         # The method should return None if the tag is not applicable
+    ...         # to the data.
+    ...
+    ...         # This checks that the Primary HDU of the data has a specific
+    ...         # keyword, 'MY_TAG'.
+    ...         if self.phu.get('MY_TAG') is not None:
[…]
