+++ _sources/examples/quickstart/index.rst.txt
@@ -1,0 +1,17 @@
+.. |file_location| replace:: ``docs/examples/quickstart/``
+
+Quickstart Example
+------------------
+
+This example contains the data and information used in our
+`Quickstart Guide <docs/quickstart.rst>`_. To go through this example, Please
+follow the steps in the `Quickstart Guide <docs/quickstart.rst>`_. This page
+only exists for the purposes of completeness in our example documentation.
+
+Files
+-----
+
+All paths are relative to the directory containing the source rst file  for
+this page (|file_location|).
+
+- ``quickstart.py``: The main script that runs the example.
