### 2024-08-02 fce2f2b
build: :heavy_plus_sign: Add devpi dependencies for build testing.
