### 2024-08-05 917012d
test(build): 🚧 Initial commit of devpi server automation.
