### 2024-08-05 bc25f88
test(build): 🚧 Initial commit of devpi server automation.
