+++ pyproject.toml
@@ -20,7 +20,6 @@ numpy = "^1.26.4"
 ruff = "^0.3.3"
 coverage = "^7.4.4"
 isort = "^5.13.2"
-poetry = "^1.8.2"
 pre-commit = "^3.6.2"
 xenon = "^0.9.1"
 
@@ -50,8 +49,7 @@ watchdog = "^4.0.1"
 
 
 [tool.poetry.group.build_test.dependencies]
-devpi-web = "^4.2.2"
-devpi-client = "^7.1.0"
+devpi = "^2.2.0"
 
 [build-system]
 requires = ["poetry-core"]
