+++ noxfile.py
@@ -15,6 +15,7 @@ TODO:
 import functools
 import subprocess
 import sys
+import time
 from pathlib import Path
 from typing import ClassVar
 
@@ -125,6 +126,181 @@ class SessionVariables:
         """Just catches accidental invocations."""
         message = "This class should not be instantiated."
         raise NotImplementedError(message)
+
+
+class DevpiServerManager:
+    """Context manager that will start and stop a devpi server."""
+
+    index_url: str | None
+    server_process: subprocess.Popen | None
+    session: nox.Session
+    tmp_dir: Path | None
+
+    active_servers: ClassVar[dict[int, "DevpiServerManager"]] = {}
+
+    # Host/port info
+    host: ClassVar[str] = SessionVariables.devpi_host
+    port: ClassVar[int] = SessionVariables.devpi_port()
+    url: ClassVar[str] = SessionVariables.devpi_url()
+
+    def __init__(
+        self, session: nox.Session, tmp_dir: Path | None = None
+    ) -> None:
+        """Initialize the context manager."""
+        self.session = session
+        self.tmp_dir = tmp_dir
+
+        # Initialize other attributes.
+        self.index_url = None
+        self.server_process = None
+
+        # Initialization methods
+        self.configure_devpi_tmp_dir()
+
+        # Register the server
+        self.active_servers[self.port] = self
+
+    def __enter__(self):
+        """Start the devpi server."""
+        session = self.session
+
+        install_test_dependencies(session, poetry_groups=["build_test"])
+
+        self.generate_config_file()
+        self.start_devpi_server()
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        """Stop the devpi server."""
+        self.stop_devpi_server()
+
+    def configure_devpi_tmp_dir(self):
+        """Configure the temporary directory for the devpi server."""
+        session = self.session
+        tmp_dir = self.tmp_dir
+
+        if tmp_dir is None:
+            temp_dir = (Path(session.create_tmp()) / "devpi").absolute()
+
+        temp_dir.mkdir()
+
+        self.tmp_dir = temp_dir
+
+    def generate_config_file(self):
+        """Generate the devpi configuration file."""
+        session = self.session
+        tmp_dir = self.tmp_dir
+        port = self.port
+
+        with session.cd(tmp_dir):
+            session.run("devpi-init", "--serverdir", ".")
+            session.run(
+                "devpi-gen-config",
+                "--serverdir",
+                tmp_dir,
+                "--port",
+                str(port),
+            )
+
+    def wait_for_devpi_startup(self, session: nox.Session) -> bool:
+        """Wait for the devpi server to start. This assumes that the server
+        has been started, there is no check for the server process itself.
+
+        It performs a curl request to the devpi server to check if it is
+        running, and it will pass for any kind of response.
+        """
+        timeout = 25
+        started = False
+
+        for check in range(1, timeout + 1):
+            session.log(f"Checking for devpi server... {check}/{timeout}")
+
+            try:
+                _ = session.run(
+                    "curl",
+                    SessionVariables.devpi_url(),
+                    silent=True,
+                    external=True,
+                )
+                started = True
+                break
+
+            except nox.command.CommandFailed:
+                time.sleep(1)
+
+        # If the process has failed, show the output and stderr
+        print(f"Server process: {self.server_process.pid}")
+
+        return started
+
+    def check_devpi_server_process(self):
+        """Check that the devpi server process is running without issue."""
+        server_process = self.server_process
+        session = self.session
+        started = self.wait_for_devpi_startup(session)
+
+        if server_process.poll() is not None or not started:
+            stdout, stderr = server_process.communicate()
[…]
