### 2024-08-06 2c8f87c
test(build): 🚧 Factor out integration/unit tests, use conda for integration tests.
