### 2024-08-06 d9de845
test(build): 🚧 Reorganize run order (switch unit and integration build tests).
