### 2024-08-07 f5b5e86
chore(dev): :art: Clean noxfile up some.
