### 2024-08-07 4a26159
chore(lint): :sparkles: Update ruff settings in pyproject.toml, remove from .pre-commit-config.yaml
