### 2024-08-07 785f0d5
ci: :art: Use nox for python/conda installation and handling for build testing.
