### 2024-08-07 22fe765
chore(lint): 🧹Lint noxfile.
