### 2024-08-07 36b8942
docs(dev): :construction: Add initial text for `nox` docs
