### 2024-08-07 4ee515b
deps(dev)!: ⏫️ Update dependencies
Add `tomli`, update `ruff`, remove `tox`
