+++ pyproject.toml
@@ -17,11 +17,12 @@ numpy = "^1.26.4"
 [tool.poetry.group.dev]
 
 [tool.poetry.group.dev.dependencies]
-ruff = "^0.3.3"
+ruff = "^0.5.6"
 coverage = "^7.4.4"
 isort = "^5.13.2"
 pre-commit = "^3.6.2"
 xenon = "^0.9.1"
+tomlkit = "^0.13.0"
 
 [tool.poetry.group.docs]
 
@@ -42,7 +43,6 @@ pytest = "^8.1.1"
 coverage = "^7.4.4"
 pytest-cov = "^4.1.0"
 pytest-doctestplus = "^1.2.1"
-tox = "^4.14.1"
 hypothesis = "^6.99.13"
 nox = "^2024.4.15"
 watchdog = "^4.0.1"
