### 2024-08-07 8afccaa
refactor(nox): :white_check_mark: Use a requirements.txt instead of manipulating poetry output.
