### 2024-08-07 69658a8
fix(nox): :adhesive_bandage: Use requirements.txt format not constraints.txt
