### 2024-08-07 b283e00
ci(lint): 📏 Ignore more rules.
