### 2024-08-07 bb934d0
fix(nox): :bug: Include main, test dependencies if none specified.
