### 2024-08-07 46710e4
chore(lint): :sparkles: Add new rules to ignore when all ruff linting rules enabled.
Right now, the code is not in need of linting with all of `ruff`'s rules. But I had some time to look through some of the rules when all are selected, and there are a numebr of useful ones in there for, e.g., documentation and catching possible bugs. This list should be treated as things that are not to be immediately implemented, or that have specific reasons/standards for beingin the code. This is on top of already-ignored rules with issues assigned.
