### 2024-08-07 ce0908b
chore(lint): ⬇️  Downgrade ruff rules to default
The rules are now deselected, I was doing some work with all rules
selected for a bit.
