### 2024-08-07 accffb6
fix(nox): Revert requirements.txt change
Commits:
  + 84bc46086f11e1809fc555ab3efc82176ee569b3
  + e4c0fb7ebd620ef326afbbc3e2fecc36666acdb7
