### 2024-08-08 277d77a
ci(dev): Re-enable build unit testing
