### 2024-08-08 0181106
feat(nox): :technologist: Add devshell and devconda sessions
