+++ noxfile.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import functools
+import os
 import socket
 import subprocess
 import sys
@@ -315,7 +316,9 @@ class DevpiServerManager:
         del self.active_servers[self.port]
 
 
-def get_poetry_dependencies(session: nox.Session, only: str = "") -> Path:
+def get_poetry_dependencies(
+    session: nox.Session, only: str = "", *, all_deps: bool = False
+) -> Path:
     """Create and return path to a requirements file from poetry.
 
     This assumes poetry is installed in the session.
@@ -328,6 +331,10 @@ def get_poetry_dependencies(session: nox.Session, only: str = "") -> Path:
     only : str, list, optional
         If provided, only return the dependencies that match the provided
         string or strings.
+
+    all : bool, optional, kw-only
+        If True, return all dependencies. Default is False. If True, the ``only``
+        argument is ignored.
 
     Returns
     -------
@@ -347,7 +354,11 @@ def get_poetry_dependencies(session: nox.Session, only: str = "") -> Path:
         "poetry",
         "show",
         f"--only={only}",
+        "--top-level",
     ]
+
+    if all_deps:
+        command.pop(2)
 
     requirements_str = session.run(
         *command,
@@ -355,9 +366,16 @@ def get_poetry_dependencies(session: nox.Session, only: str = "") -> Path:
         silent=True,
     )
 
+    session.run("python", "--version", silent=True)
+
     requirements_str = requirements_str.replace("(!)", "   ").strip()
 
-    lines = [line.split()[0].strip() for line in requirements_str.split("\n")]
+    lines = [
+        line.split()[0].strip()
+        for line in requirements_str.split("\n")
+        if not line.startswith("Skipping virtualenv creation")
+    ]
+
     req_file_path.write_text("\n".join(lines))
 
     return req_file_path
@@ -750,6 +768,142 @@ def linting(session: nox.Session) -> None:
 
     # Run the linters.
     session.run("ruff", "lint", ".")
+
+
+@nox.session(venv_backend="none")
+def devshell(session: nox.Session) -> None:
+    """Create a venv for development."""
+    session.env["POETRY_PREFER_ACTIVE_PYTHON"] = "true"
+    session.env["POETRY_VIRTUALENVS_CREATE"] = "false"
+
+    venv_path = Path(".astrodata_venv/").absolute()
+    activate_path = venv_path / "bin" / "activate"
+
+    # Check that poetry is installed
+    try:
+        _ = session.run("poetry", "--version", silent=True)
+
+    except nox.command.CommandFailed as err:
+        message_lines = (
+            "Poetry is not installed. Please install poetry before running ",
+            "this session.",
+            "Installation instructions can be found at ",
+            "https://python-poetry.org/docs/#installation.",
+        )
+        message = "\n".join(message_lines)
+
+        raise RuntimeError(message) from err
+
+    # Remove any existing venv
+    if venv_path.exists():
+        session.run("rm", "-rf", str(venv_path))
+
+    # Create the venv
+    _ = session.run(
+        "python", "-m", "venv", str(venv_path), "--prompt", "astrodata_venv"
+    )
+
+    req_file_path = get_poetry_dependencies(session, all_deps=True)
+    venv_python_bin = venv_path / "bin" / "python"
+
+    _ = session.run(
+        str(venv_python_bin), "-m", "pip", "install", "--upgrade", "pip"
+    )
+
+    _ = session.run(
+        str(venv_python_bin),
+        "-m",
+        "pip",
+        "install",
+        "--requirement",
+        str(req_file_path),
+    )
+
+    session.log("Virtual environment created.")
[…]
