### 2024-08-08 96b97f9
fix(nox): 🧹 Remove assignments to throwaway vars in noxfile.
