+++ noxfile.py
@@ -216,7 +216,7 @@ class DevpiServerManager:
             session.log(f"Checking for devpi server... {check}/{timeout}")
 
             try:
-                _ = session.run(
+                session.run(
                     "curl",
                     SessionVariables.devpi_url(),
                     silent=True,
@@ -480,7 +480,7 @@ def dragons_release_tests(session: nox.Session) -> None:
     # Positional arguments after -- are passed to pytest.
     pos_args = session.posargs
 
-    _ = session.run(
+    session.run(
         "pytest",
         *SessionVariables.dragons_pytest_options,
         *pos_args,
@@ -554,7 +554,7 @@ def dragons_dev_tests(session: nox.Session) -> None:
     # Positional arguments after -- are passed to pytest.
     pos_args = session.posargs
 
-    _ = session.run(
+    session.run(
         "pytest",
         *SessionVariables.dragons_pytest_options,
         *pos_args,
@@ -571,14 +571,14 @@ def unit_tests(session: nox.Session) -> None:
     pos_args = session.posargs
 
     # Run the tests. Need to pass arguments to pytest.
-    _ = session.run("pytest", *SessionVariables.unit_pytest_options, *pos_args)
+    session.run("pytest", *SessionVariables.unit_pytest_options, *pos_args)
 
 
 @nox.session(venv_backend="conda", python=SessionVariables.python_versions)
 def conda_unit_tests(session: nox.Session) -> None:
     """Run the unit tests."""
     # Configure session channels.
-    _ = session.run(
+    session.run(
         "conda",
         "config",
         "--env",
@@ -593,7 +593,7 @@ def conda_unit_tests(session: nox.Session) -> None:
     pos_args = session.posargs
 
     # Run the tests. Need to pass arguments to pytest.
-    _ = session.run("pytest", *SessionVariables.unit_pytest_options, *pos_args)
+    session.run("pytest", *SessionVariables.unit_pytest_options, *pos_args)
 
 
 def unit_test_build(session: nox.Session) -> None:
@@ -613,7 +613,7 @@ def unit_test_build(session: nox.Session) -> None:
     pos_args = session.posargs
 
     # Run the tests. Need to pass arguments to pytest.
-    _ = session.run("pytest", *SessionVariables.unit_pytest_options, *pos_args)
+    session.run("pytest", *SessionVariables.unit_pytest_options, *pos_args)
 
 
 def integration_test_build(session: nox.Session) -> None:
@@ -642,7 +642,7 @@ def integration_test_build(session: nox.Session) -> None:
     # Positional arguments after -- are passed to pytest.
     pos_args = session.posargs
 
-    _ = session.run(
+    session.run(
         "pytest",
         *SessionVariables.dragons_pytest_options,
         *pos_args,
@@ -656,10 +656,10 @@ def coverage(session: nox.Session) -> None:
     install_test_dependencies(session)
 
     # Generate the coverage report.
-    _ = session.run("coverage", "report", "--show-missing")
+    session.run("coverage", "report", "--show-missing")
 
     # Generate the HTML report.
-    _ = session.run("coverage", "html")
+    session.run("coverage", "html")
 
 
 @nox.session
@@ -673,8 +673,8 @@ def docs(session: nox.Session) -> None:
     # Build the documentation.
     # TODO(teald): Add nitpicky flag to fix warnings. -- Issue #42
     target = Path("_build")
-    _ = session.run("rm", "-rf", target, external=True)
-    _ = session.run("sphinx-build", "docs", target)
+    session.run("rm", "-rf", target, external=True)
+    session.run("sphinx-build", "docs", target)
 
 
 def use_devpi_server(func):
[…]
