+++ docs/developer/testing.rst
@@ -32,7 +32,7 @@ These tests include:
   behaves as expected when different parts of the codebase are combined.
 - **Script tests**: These tests are used to test the scripts in the codebase.
   They are used to ensure that the scripts behave as expected when run. these
-  primarily test example scripts within teh documentation.
+  primarily test example scripts within the documentation.
 - **Build/Release tests**: These tests are used to test the build and release
   process, and the integrity of the package being released. They are used to
   ensure that the code can be built and released correctly.
@@ -44,8 +44,163 @@ make understanding and interpretting test results easier.
 Running Tests
 =============
 
-.. |developer install guide| replace:: :doc:`developer/index.rst`
+.. |developer install guide| replace:: :doc:`index`
 
 Running the tests involves some setup. It's assumed you have already installed
 and setup |Poetry| as per the |developer install guide|, or are prepared to
 work with the |astrodata| source code in some other way.
+
+If you'd like to run the test in your local environment, you can do so by
+running the following command from the root of the repository:
+
+.. code-block:: bash
+
+    pytest
+
+This is discouraged, but can be useful for runnin spcific tests or debugging.
+To Isolate a test, use a keyword and/or point to a file:
+
+.. code-block:: bash
+
+    pytest -k "test_foo" tests/test_foo.py
+
+The above command will run all tests that match the keyword "test_foo" in their
+names in the file ``tests/test_foo.py``.
+
+
+|nox| is the preferred way to run tests, as it ensures that the tests are run in
+clean and isolated environments. Simply run:
+
+.. code-block:: bash
+
+    nox
+
+This will run a subset of the tests recommended for local development, the unit
+tests, and their coverage. To see all possible tests, run:
+
+.. code-block:: bash
+
+    nox -l
+
+.. _nox_documentation: https://nox.thea.codes/en/stable/usage.html
+
+This will output information about sessions you can select with ``nox -s``. For
+more information about using |nox|, see the `Nox CLI Documentation
+<nox_documentation>`_.
+
+.. note::
+    |nox| is also used for other automation tasks, such as:
+      + building the documentation (``nox -s docs``)
+      + linting & formatting (``nox -s linting``)
+      + creating development environments (``nox -s devshell``)
+
+Writing Tests
+=============
+
+Writing tests is an important part of contributing to |astrodata|. Tests help
+ensure that the code behaves as expected, and that changes to the code don't
+break existing functionality.
+
+When writing tests, it's important to follow the `testing best practices
+<https://docs.pytest.org/en/stable/goodpractices.html>`_ outlined in the
+|pytest|_.
+
+Tests are located in the ``tests/`` directory, and are organized by the
+type of test they are. For example, unit tests are located in the
+``tests/unit/`` directory, integration tests are located in the
+``tests/integration/`` directory, and so on.
+
+Unit Tests
+----------
+
+Unit tests are used to test individual units of code, such as functions or
+classes. They are used to ensure that the code behaves as expected when given
+specific inputs. They are also the most common type of test in |astrodata|,
+and likely the most common type of test you will write.
+
+Unit tests are in the ``tests/unit`` directory. Testing for specific modules
+are collected in individual files. For example, a test for a function in the
+``astrodata/utils.py`` module would be located in ``tests/unit/test_utils.py``.
+
+You can run all unit tests (and nothing else) with |nox| by running:
+
+.. code-block:: bash
+
+    nox -s unit
+
+Unit tests require that coverage increases if new code is added. It is highly
+encouraged to write tests for contributions before writing new code. If you get
+stuck on how to test your idea, feel free to ask for help in the issue or pull
[…]
