+++ docs/conf.py
@@ -175,6 +175,7 @@ rst_prolog = """
    :format: html
 
 .. |Poetry| replace:: `Poetry <https://python-poetry.org/>`__
+.. |nox| replace:: `nox <https://nox.thea.codes/en/stable/>`__
 .. |DRAGONS| replace:: `DRAGONS <https://dragons.readthedocs.io/>`__
 .. |RSProgManual| replace:: `Recipe System Programmer Manual <http://dragons-recipe-system-programmers-manual.readthedocs.io/en/{v}/>`__
 .. |RSUserManual| replace:: `Recipe System User Manual <http://dragons-recipe-system-users-manual.readthedocs.io/en/{v}/>`__
