+++ docs/developer/index.rst
@@ -18,7 +18,10 @@ running the unit tests.
    :maxdepth: 2
    :caption: Contents:
 
+   documentation
    testing
+
+.. _dev_environment_setup:
 
 Developer environment setup
 ===========================
@@ -78,26 +81,44 @@ Install the dependencies
    cd astrodata
 
 If you are not already inside a virtual environment of some flavor, such as a
-``conda`` or ``venv`` environment, create one now. For example, to create a new
-virtual environment with ``venv``:
+``conda`` or ``venv`` environment, you can have ``nox`` create one for you:
 
 .. code-block:: bash
 
-   python -m venv .venv
-   source .venv/bin/activate
+   nox -s devshell  # Creates a virtual environment
+   nox -s devconda  # Creates a conda environment
 
-To install the dependencies, navigate to the root of the repository and run:
+.. note::
+
+   We recommend using the ``devshell`` session, as ``venv`` environments are
+   more flexible and allow for more pythonic automation in the development
+   process. The ``devconda`` session is provided for users who prefer to use
+   conda environments, but it's not as flexible as the ``devshell`` session and
+   may require manual tweaking.
+
+.. warning::
+
+   Both of these sessions will create a new environment, either located at
+   ``.astrodata_venv`` or in the ``astrodata`` conda environment. It will
+   remove any environments found there (for consistency when resetting the
+   environment).
+
+Otherwise, to install the dependencies in the current environment--or one
+you've set up and activate yourself--navigate to the root of the repository and
+run:
 
 .. code-block:: bash
 
    poetry install
 
 
+.. _editable_pip: https://setuptools.pypa.io/en/latest/userguide/development_mode.html
+
 This will install all the dependencies needed to run |astrodata| within your
 virtual environment, including all test, development, and documentation
 dependencies. This installs |astrodata| in a way that's equivalent to
 installing with
-`pip in editable mode <https://setuptools.pypa.io/en/latest/userguide/development_mode.html>`.
+`pip in editable mode <editable_pip>`.
 
 .. note::
     You can install specific dependency groups by running ``poetry install
@@ -118,6 +139,7 @@ Run the tests
 -------------
 
 .. _nox: https://nox.thea.codes/en/stable/
+.. |nox| replace:: ``nox``
 
 |astrodata| uses nox_ for running tests. To run the tests, simply run:
 
@@ -146,47 +168,18 @@ Or, to run the normal unit tests and not the linter:
 
 .. code-block:: bash
 
-   nox -s "unit"
+   nox -s "unit_tests"
 
 All tests will be run in isolated environments based on specifications in the
 ``noxfile.py`` file in the main project directory. Those environments, by
 default, are re-created each time you run the tests. To avoid that, you can
-pass the ``--reuse-existing-virtualenvs``/``-r`` flag to ``nox``, which will
+pass the ``--reuse-existing-virtualenvs``/``-r`` flag to |nox|, which will
 reuse any existing virtual environments it finds. If the environment isn't
 found, it will be made.
 
 
 Other development commands
 --------------------------
-
-
-Development with a Poetry environment
-=====================================
-
-|Poetry| also has a feature to create a shell with the dependencies installed.
-This is useful for development, as it allows you to run commands in the
-environment without activating it. To create a shell, run:
-
-.. code-block:: bash
-
-   poetry shell
-
-This will create a shell with the dependencies installed. You can then run
-commands in this shell as you would in a normal shell. To exit the shell, run:
-
-.. code-block:: bash
-
-   exit
-
-This takes similar steps to the above, but make Poetry handle the environment
-for you. While this is convenient, it can be confusing if you're not familiar
-with virtual environments and the shell command itself is somewhat limited in
[…]
