### 2024-08-09 4c64fd4
ci(build)👷 Separated build test jobs into different runs.
