### 2024-08-09 2c918bd
feat(ci): ⚒️ Add conventional pre-commit
Now, when a message is added to a commit it is checked for consistency
with the Conventional Commit standards. See their page:

+ [Conventional Commit
  Standard](https://www.conventionalcommits.org/en/v1.0.0/)
