+++ .pre-commit-config.yaml
@@ -42,3 +42,10 @@ repos:
               build|
               dist|
               docs"""
+
+- repo: https://github.com/compilerla/conventional-pre-commit
+  rev: v3.4.0
+  hooks:
+    - id: conventional-pre-commit
+      stages: [commit-msg]
+      args: []
