### 2024-08-09 5f9f7ae
docs(README): ✍️ Add build/release test badge
