+++ README.md
@@ -12,7 +12,7 @@
 [pypi package version badge]: https://badge.fury.io/py/astrodata.svg
 [unit test status badge]: https://github.com/GeminiDRSoftware/astrodata/actions/workflows/unit_tests.yml/badge.svg
 [integration test status badge]: https://github.com/GeminiDRSoftware/astrodata/actions/workflows/integration_tests.yml/badge.svg
-
+[build test status badge]: https://github.com/GeminiDRSoftware/astrodata/actions/workflows/build_tests.yml/badge.svg
 `astrodata`
 =============
 
@@ -35,6 +35,7 @@
 ![A badge displaying the testing coverage percentage of this repository.][coverage badge]
 ![Unit test status badge][unit test status badge]
 ![Integration test status badge][integration test status badge]
+![Build/Release test status badge][build test status badge]
 
 ### Building & Publishing
 ![Documentation build status badge][docs build badge]
