### 2024-08-09 ac88bf4
feat(nox): 👷 Add pre-commit init session
