### 2024-08-13 c29df6e
fix(nox): ⏫️ Use poetry export for nox reqs
