### 2024-08-13 058d6c9
test(scripts): 🚧 Add initial script testing
