+++ tests/script_tests/conftest.py
@@ -1,0 +1,1 @@
+"""Configuration for pytest script tests."""
