### 2024-08-13 2a5706b
test(scripts): 🚧 Add initial script testing
