### 2024-08-13 a5d5993
test(lint): 🚧 Update linting settings to capture new testing code.
