### 2024-08-13 a70b875
test(lint): 🚧 Update linting settings to capture new testing code.
