### 2024-08-13 ea13a18
fix(pre-commit): 🐛 Remove old formatting args from ruff format
