### 2024-08-13 e6c3860
test(build): :construction: Add build session, working script test.
