### 2024-08-13 0d40b09
ci(scripts): 🚧 Add script test workflow.
