### 2024-08-13 70c7a8c
ci(scripts): 🚧 Add warning handling for script tests
