### 2024-08-13 09381b2
chore(lint): 🧹 Lint by hand (doc/strings)
