### 2024-08-13 9c567b8
chore(lint): 🧹 Lint by hand (doc/strings)
