### 2024-08-13 2f01cd5
chore(nox): 🧹 Remove unecessary variable assignment
