### 2024-08-13 7c96787
ci(tests)!: 🎨 Consolidate testing workflows
This commit consolidates what was onces a number of testing workflows
for source code testing into a single file (like build tests). This
reduces spaminess with emails/notifications while still performing all
the tests.

BREAKING CHANGE: The unit_test.yml, documentation_tests.yml, and
integration_tests.yml files are no longer in the main workflows folder.
