### 2024-08-13 909b5fe
ci(test): 🧹 Fix typo, update names
