### 2024-08-13 c44013b
ci(test): 🧹 Remove windows test accidentally included.
