+++ .github/workflows/source_tests.yml
@@ -1,4 +1,4 @@
-name: "Testing"
+name: "Source testing"
 
 on:
     push:
@@ -76,7 +76,6 @@ jobs:
                 os:
                     - ubuntu-latest
                     - macos-latest
-                    - windows-latest
 
         steps:
             - name: "Check out the repo"
