### 2024-08-13 5067f26
docs: 🎨 Remove doctest formatting, statement output.
