### 2024-08-13 fb6f7fd
test(scripts): 🧪 Add usermanual intro to tests
