+++ pyproject.toml
@@ -217,3 +217,4 @@ mark-parentheses = false
 
 [tool.script_tests]
 "docs/quickstart.rst" = ""
+"docs/manuals/usermanual/intro.rst" = ""
