### 2024-08-13 1e0d3eb
ci(test,scripts): 🚧 Add build tests with scripts
