### 2024-08-13 854a1bc
ci: ✍️ Specify build in build test names.
