+++ .github/workflows/build_tests.yml
@@ -15,7 +15,7 @@ defaults:
 
 jobs:
     unit_tests:
-        name: "Testing on ${{ matrix.os }}"
+        name: "Build testing on ${{ matrix.os }}"
         runs-on: "${{ matrix.os }}"
 
         strategy:
@@ -47,7 +47,7 @@ jobs:
 
 
     documentation_tests:
-        name: "Documentation testing on ${{ matrix.os }}"
+        name: "Build documentation testing on ${{ matrix.os }}"
         runs-on: "${{ matrix.os }}"
 
         strategy:
@@ -93,7 +93,7 @@ jobs:
 
 
     integration_tests:
-        name: "Testing on ${{ matrix.os }}"
+        name: "Build integration testing on ${{ matrix.os }}"
         runs-on: "${{ matrix.os }}"
 
         strategy:
