### 2024-08-13 0d3a623
ci: ✍️ Use normal python setup for build tests.
