+++ .github/workflows/build_tests.yml
@@ -15,7 +15,7 @@ defaults:
 
 jobs:
     unit_tests:
-        name: "Build testing on ${{ matrix.os }}"
+        name: "Build unit testing on ${{ matrix.os }}"
         runs-on: "${{ matrix.os }}"
 
         strategy:
@@ -25,22 +25,22 @@ jobs:
                     - ubuntu-latest
                     - macos-latest
 
-                python-version:
-                    - "3.10"
-                    - "3.11"
-                    - "3.12"
-
         steps:
-            - uses: "actions/checkout@v4"
-            - uses: excitedleigh/setup-nox@v2.0.0
-            - uses: conda-incubator/setup-miniconda@v3
-              with:
-                python-version: "${{ matrix.python-version }}"
+            - name: "Check out the repo"
+              uses: "actions/checkout@v4"
 
-            - name: "Install and update pip"
+            - name: "Set up Python"
+              uses: "actions/setup-python@v5"
+              with:
+                  python-version: |
+                    3.10
+                    3.11
+                    3.12
+
+            - name: "Install dependencies"
               run: |
-                  conda install -c conda-forge pip
-                  pip install --upgrade pip poetry
+                  python -m pip install --upgrade pip
+                  python -m pip install nox poetry
 
             - name: "Unit tests"
               run: nox -s build_tests_unit -x -- -vv
@@ -89,7 +89,7 @@ jobs:
             - name: "Run nox"
               run: |
                   # Unit tests
-                  python -m nox -s build_tests_scripts --verbose -x -- -xvv
+                  python -m nox -s build_tests_scripts --verbose -x -- -vv
 
 
     integration_tests:
