### 2024-08-13 7a70a3f
docs: :memo: Remove dollar sign, fix clone URL in intro
