### 2024-08-13 eaa87cc
fix(nox): 👷 Fix issue with pre-commit, astrodata install
