### 2024-08-13 a334994
refactor(package): :art: Add testing to top-level imports.
This addresses an issue with testing not being available. It's currently the best way to make succinct examples of astrodata concepts, and importing it doesn't bring on any downsides that I can see beyond users potentially being confused about it (which is fixable with refactoring it and updating documentation).
