### 2024-08-13 e2ea3b4
fix(nox,lint): 🐛 Fix incorrect command in nox lint session
