+++ docs/examples/gemini_examples/index.rst
@@ -1,15 +1,0 @@
-.. _gemini_examples:
-
-Gemini Examples
-===============
-
-These examples use the |Gemini| |DRAGONS| data reduction pipeline. They use
-external code to perform operations outside of |astrodata|'s capabilities.
-
-Specifically, |DRAGONS| extends |AstroData| in its |gemini_instruments|
-package. This package is a grounded example of how to extend |AstroData| for
-specific instruments. All reduction functionality is supported by other parts
-of the library.
-
-To use these examples, you need to install |DRAGONS| and its dependencies.
-Please follow the |DRAGONS| installation instructions at |DRAGONS_install|.
