### 2024-08-14 b372694
chore(docs): 🧹 Remove outdated examples section.
