+++ docs/examples/index.rst
@@ -1,16 +1,0 @@
-.. _examples_landing_page:
-
-Examples
-========
-
-We provide a number of examples to aid in understanding specific use cases,
-common patterns, and best practices when working with |astrodata|. To see a
-complete user guide, please see our |UserGuide|. For a deeper guide to
-programming with |astrodata|, please visit our |DeveloperGuide|.
-
-.. toctree::
-    :maxdepth: 2
-
-    generic_examples/index.rst
-    gemini_examples/index.rst
-    quickstart/index.rst
