+++ docs/examples/gemini_examples/astrodata_gemini_examples.yml
@@ -1,219 +1,0 @@
-name: astrodata_gemini_examples
-channels:
-  - http://astroconda.gemini.edu/public
-  - conda-forge
-  - defaults
-dependencies:
-  - aom=3.8.1=h73e2aa4_0
-  - asdf=3.0.1=pyhd8ed1ab_0
-  - asdf-astropy=0.5.0=pyhd8ed1ab_0
-  - asdf-coordinates-schemas=0.2.0=pyhd8ed1ab_1
-  - asdf-standard=1.0.3=pyhd8ed1ab_0
-  - asdf-transform-schemas=0.4.0=pyhd8ed1ab_0
-  - asdf-unit-schemas=0.1.0=pyhd8ed1ab_0
-  - asdf-wcs-schemas=0.3.0=pyhd8ed1ab_0
-  - astropy=6.0.0=py310h91862f5_0
-  - astropy-iers-data=0.2024.2.12.0.31.0=pyhd8ed1ab_0
-  - astroquery=0.4.6=pyhd8ed1ab_0
-  - astroscrappy=1.1.0=py310h936d966_2
-  - attrs=23.2.0=pyh71513ae_0
-  - backports.zoneinfo=0.2.1=py310h2ec42d9_8
-  - beautifulsoup4=4.12.3=pyha770c72_0
-  - bleach=6.1.0=pyhd8ed1ab_0
-  - blosc=1.21.5=heccf04b_0
-  - bokeh=2.4.3=pyhd8ed1ab_3
-  - bottleneck=1.3.7=py310hf0b6da5_1
-  - brotli=1.1.0=h0dc2134_1
-  - brotli-bin=1.1.0=h0dc2134_1
-  - brotli-python=1.1.0=py310h9e9d8ca_1
-  - brunsli=0.1=h046ec9c_0
-  - bzip2=1.0.8=h1de35cc_0
-  - c-ares=1.26.0=h10d778d_0
-  - c-blosc2=2.13.2=h0ae8482_0
-  - ca-certificates=2024.2.2=h8857fd0_0
-  - certifi=2024.2.2=pyhd8ed1ab_0
-  - cfitsio=4.3.1=h60fb419_0
-  - charls=2.4.2=he965462_0
-  - charset-normalizer=3.3.2=pyhd8ed1ab_0
-  - click=8.1.7=unix_pyh707e725_0
-  - colorama=0.4.6=pyhd8ed1ab_0
-  - colorcet=3.0.1=pyhd8ed1ab_0
-  - contourpy=1.2.0=py310ha697434_0
-  - coverage=7.4.1=py310hb372a2b_0
-  - cycler=0.12.1=pyhd8ed1ab_0
-  - dav1d=1.2.1=h0dc2134_0
-  - dragons=3.1.0=py310_2
-  - ds9=8.2.1=0
-  - exceptiongroup=1.2.0=pyhd8ed1ab_2
-  - expat=2.5.0=hf0c8a7f_1
-  - fftw=3.3.10=nompi_h4fa670e_108
-  - fitsverify=4.22=hf2275c7_0
-  - font-ttf-dejavu-sans-mono=2.37=hab24e00_0
-  - font-ttf-inconsolata=3.000=h77eed37_0
-  - font-ttf-source-code-pro=2.038=h77eed37_0
-  - font-ttf-ubuntu=0.83=h77eed37_1
-  - fontconfig=2.14.2=h5bb23bf_0
-  - fonts-conda-ecosystem=1=0
-  - fonts-conda-forge=1=0
-  - fonttools=4.48.1=py310hb372a2b_0
[…]
