### 2024-08-14 212254e
docs(conf): 🧹 Remove reference to example sesction in conf.py
