+++ docs/conf.py
@@ -138,7 +138,6 @@ rst_prolog = """
 .. |UserManual| replace:: |UserGuide|
 .. |ProgManual| replace:: |DeveloperGuide|
 .. |ProgrammerManual| replace:: |DeveloperGuide|
-.. |Examples| replace:: :doc:`Examples </examples/index>`
 .. |QuickStart| replace:: :doc:`Quick Start </quickstart>`
 .. |DeveloperInstall| replace:: :doc:`Developer Installation </developer/index>`
 
