### 2024-08-14 941cd02
docs(nox,examples): Remove example references
This removes all references to the examples section made through links.
It also sets nox to raise sphinx warnings as errors (using `sphinx-build
-W`).
