### 2024-08-14 99435fa
refactor: :wastebasket: Deprecate "add_header_to_table"
