### 2024-08-14 405e7f2
docs: ✍️ State deprecation of add_header_to_table in summary
