### 2024-08-14 25a0ea2
docs: ✍️ Update docstring for descriptor decorator.
