+++ astrodata/utils.py
@@ -174,7 +174,7 @@ class TagSet(namedtuple("TagSet", "add remove blocked_by blocks if_present")):
 
 
 def astro_data_descriptor(fn):
-    """Decorator that will mark a class method as an AstroData descriptor.
+    """Mark a class method as an AstroData descriptor.
 
     Args
     -----
@@ -288,13 +288,7 @@ def assign_only_single_slice(fn):
 
 
 def astro_data_tag(fn):
-    """Decorator that marks methods of an `AstroData` derived class as part of
-    the tag-producing system.
-
-    It wraps the method around a function that will ensure a consistent return
-    value: the wrapped method can return any sequence of sequences of strings,
-    and they will be converted to a TagSet. If the wrapped method
-    returns None, it will be turned into an empty TagSet.
+    """Mark a method as a tag-producing method.
 
     Args
     -----
@@ -304,6 +298,31 @@ def astro_data_tag(fn):
     Returns
     --------
     A wrapper function
+
+    Notes
+    -----
+    Decorator that marks methods of an `AstroData` derived class as part of
+    the tag-producing system.
+
+    It wraps the method around a function that will ensure a consistent return
+    value: the wrapped method can return any sequence of sequences of strings,
+    and they will be converted to a TagSet. If the wrapped method
+    returns None, it will be turned into an empty TagSet.
+
+    Example
+    -------
+
+    .. code-block:: python
+
+        class MyAstroData(AstroData):
+            @astro_data_tag
+            def my_tag_method(self):
+                # Below are the tags generated by this method based on whether
+                # the instrument is GMOS or not.
+                if self.phu.get('INSTRUME') == 'GMOS':
+                    return {'GMOS'}
+
+                return {'NOT_GMOS'}
     """
 
     @wraps(fn)
