### 2024-08-14 5428ee9
docs: ✍️ Update returns_list documentation.
