### 2024-08-14 11d7390
docs(api): ✍️ Update from_file documentation
