+++ astrodata/__init__.py
@@ -64,8 +64,44 @@ factory.add_class(AstroData)
 def from_file(*args, **kwargs):
     """Return an |AstroData| object from a file.
 
+    Arguments
+    ---------
+    source: str, os.PathLike or HDUList
+        The path to the file or the HDUList object. If a string is passed, it
+        will be treated as a path to a file.
+
+    Returns
+    -------
+    AstroData
+        An instantiated object. It will be a subclass of |AstroData|.
+
+    Notes
+    -----
     For implementation details, see
-    :meth:`~astrodata.AstroDataFactory.get_astro_data`.
+    :py:meth:`~astrodata.AstroDataFactory.get_astro_data`.
+
+    This function is a wrapper around the factory method
+    :py:meth:`~astrodata.AstroDataFactory.get_astro_data`, and uses the
+    default factory instance at :py:attr:`~astrodata.factory`. If you want to
+    override the default factory, you can create a new instance of
+    :py:class:`~astrodata.AstroDataFactory` and use its methods directly, or
+    assign it to :py:attr:`~astrodata.factory`.
+
+    Example
+    -------
+
+    >>> from astrodata import from_file
+    >>> ad = from_file("path/to/file.fits")
+
+    Alternatively, you can use an :py:class:`~astropy.io.fits.HDUList` object:
+
+    >>> from astropy.io import fits
+    >>> hdulist = fits.open("path/to/file.fits")
+    >>> ad = from_file(hdulist)
+
+    Which can be useful for inspecting input before creating the |AstroData|
+    object. This will not use the normal |AstroData| lazy-loading mechanism,
+    however.
     """
     return factory.get_astro_data(*args, **kwargs)
 
@@ -73,8 +109,30 @@ def from_file(*args, **kwargs):
 def create(*args, **kwargs):
     """Return an |AstroData| object from data.
 
-    For implementation details, see
-    :meth:`~astrodata.AstroDataFactory.create_from_scratch`
+    Arguments
+    ---------
+    phu : `fits.PrimaryHDU` or `fits.Header` or `dict` or `list`
+        FITS primary HDU or header, or something that can be used to create
+        a fits.Header (a dict, a list of "cards").
+
+    extensions : list of HDUs
+        List of HDU objects.
+
+    Returns
+    -------
+    `astrodata.AstroData`
+        An AstroData instance.
+
+    Raises
+    ------
+    ValueError
+        If ``phu`` is not a valid object.
+
+    Example
+    -------
+
+    >>> from astrodata import create
+    >>> ad = create(phu=fits.PrimaryHDU(), extensions=[fits.ImageHDU()])
     """
     return factory.create_from_scratch(*args, **kwargs)
 
