### 2024-08-14 a94c37a
docs(api): ✍️ Update open docstring
