### 2024-08-14 GeminiDRSoftware/astrodata#16 @teald
This issue is ongoing during some service outages with the gemini archive.
