### 2024-08-15 aa5d6de
docs: ✍️ Unify and expand on docstrings in AstoData
