+++ astrodata/core.py
@@ -42,8 +42,8 @@ NO_DEFAULT = object()
 _ARIT_DOC = """
     Performs {name} by evaluating ``self {op} operand``.
 
-    Parameters
-    ----------
+    Arguments
+    ---------
     oper : number or object
         The operand to perform the operation  ``self {op} operand``.
 
@@ -54,11 +54,12 @@ _ARIT_DOC = """
 
 
 class AstroData:
-    """Base class for the AstroData software package. It provides an interface
-    to manipulate astronomical data sets.
+    """Base class for the AstroData software package.
 
-    Parameters
-    ----------
+    It provides an interface to manipulate astronomical data sets.
+
+    Arguments
+    ---------
     nddata : `astrodata.NDAstroData` or list of `astrodata.NDAstroData`
         List of NDAstroData objects.
 
@@ -73,7 +74,51 @@ class AstroData:
         object will access to. This is used when slicing an object, then the
         sliced AstroData will have the ``.nddata`` list from its parent and
         access the sliced NDAstroData through this list of indices.
+
+
+    .. warning::
+
+        This class is not meant to be instantiated directly. Instead, use the
+        factory method :py:func:`astrodata.from_file` to create an instance of this
+        class using a file. Alternatively, use the :py:meth:`astrodata.create`
+        function to create a new instance from scratch.
+
+        The documentation here is meant for developers who want to extend the
+        functionality of this class.
+
+    Registering an |AstroData| subclass
+    -----------------------------------
+
+    To create a new subclass of |AstroData|, you need to register it with the
+    factory. This is done by creating a new subclass and using
+    :py:meth:`AstroDataFactory.add_class`:
+
+    .. code-block:: python
+
+        from astrodata import AstroData, factory
+
+        class MyAstroData(AstroData):
+            @classmethod
+            def _matches_data(cls):
+                '''Trivial match for now.'''
+                return True
+
+        factory.add_class(MyAstroData)
+
+    Once the class is registered, the factory will be able to create instances
+    of it when reading files. It will also be able to create instances of it
+    when using the :py:meth:`astrodata.create` function. It uses the special,
+    required method :py:meth:`~astrodata.AstroData._matches_data` to determine
+    if the class matches the data in the file. If there are multiple matches,
+    the factory will try to find one that has the most specific match and is
+    a subclass of the other candidates.
+
+    If There is no match or multiple matches, the factory will raise an
+    exception. See :py:meth:`AstroDataFactory.get_astro_data` for more
+    information.
     """
+
+    # TODO(teald): Docstring for AstroData has a bad lin to AstroDataFactory
 
     # Derived classes may provide their own __keyword_dict. Being a private
     # variable, each class will preserve its own, and there's no risk of
@@ -139,10 +184,10 @@ class AstroData:
         self._path = None
 
     def __deepcopy__(self, memo):
-        """Returns a new instance of this class.
+        """Return a new instance of this class.
 
-        Parameters
-        ----------
+        Arguments
+        ---------
         memo : dict
             See the documentation on `deepcopy` for an explanation on how
             this works.
@@ -157,10 +202,10 @@ class AstroData:
         return obj
 
     def _keyword_for(self, name):
-        """Returns the FITS keyword name associated to ``name``.
+        """Return the FITS keyword name associated to ``name``.
 
-        Parameters
-        ----------
+        Arguments
+        ---------
         name : str
             The common "key" name for which we want to know the associated
             FITS keyword.
@@ -232,10 +277,10 @@ class AstroData:
 
     @classmethod
     def matches_data(cls, source) -> bool:
-        """Returns True if the class can handle the data in the source.
+        """Return True if the class can handle the data in the source.
 
-        Parameters
-        ----------
+        Arguments
+        ---------
         source : list of `astropy.io.fits.HDUList`
[…]
