### 2024-08-15 601fdc3
docs: ✍️ Unify and update nddata.py docs
