+++ astrodata/nddata.py
@@ -39,7 +39,9 @@ class ADVarianceUncertainty(VarianceUncertainty):
 
 
 class AstroDataMixin:
-    """A Mixin for ``NDData``-like classes (such as ``Spectrum1D``) to enable
+    """Mixin with AstroData-like behavior for NDData-like classes.
+
+    Mixin for ``NDData``-like classes (such as ``Spectrum1D``) to enable
     them to behave similarly to ``AstroData`` objects.
 
     These behaviors are:
@@ -53,8 +55,11 @@ class AstroDataMixin:
     """
 
     def __getattr__(self, attribute):
-        """Allow access to attributes stored in self.meta['other'], as we do
-        with AstroData objects.
+        """Access attributes stored in self.meta['other'].
+
+        Required to access attributes like |AstroData| objects. See the
+        documentation for |AstroData|'s ``__getattr__`` method for more
+        information.
         """
         if attribute.isupper():
             try:
@@ -80,8 +85,14 @@ class AstroDataMixin:
         compare_wcs="first_found",
         **kwds,
     ):
-        """Override the NDData method so that "bitwise_or" becomes the default
+        """Perform arithmetic operations on the data.
+
+        Overrides the NDData method so that "bitwise_or" becomes the default
         operation to combine masks, rather than "logical_or"
+
+        .. warning::
+            This method is not intended to be called directly. Use the
+            arithmetic methods of the NDData object instead.
         """
         return super()._arithmetic(
             operation,
@@ -95,9 +106,16 @@ class AstroDataMixin:
         )
 
     def _slice_wcs(self, slices):
-        """The ``__call__()`` method of gWCS doesn't appear to conform to the
+        """Slice the WCS object.
+
+        The ``__call__()`` method of gWCS doesn't appear to conform to the
         APE 14 interface for WCS implementations, and doesn't react to slicing
         properly. We override NDSlicing's method to do what we want.
+
+        Arguments
+        ---------
+        slices : slice or tuple of slices
+            The slice or slices to apply to the WCS.
         """
         if not isinstance(self.wcs, gWCS):
             return self.wcs
@@ -183,7 +201,7 @@ class AstroDataMixin:
 
     @property
     def variance(self):
-        """A convenience property to access the contents of ``uncertainty``."""
+        """Access the contents of ``uncertainty``."""
         return getattr(self.uncertainty, "array", None)
 
     @variance.setter
@@ -194,7 +212,9 @@ class AstroDataMixin:
 
     @property
     def wcs(self):
-        """The WCS of the data. This is a gWCS object, not a FITS WCS object.
+        """Return the WCS of the data as a gWCS object.
+
+        This is a gWCS object, not a FITS WCS object.
 
         This is returning wcs from an inhertited class, see NDData.wcs for more
         details.
@@ -249,7 +269,7 @@ class NDWindowing:
 class NDWindowingAstroData(
     AstroDataMixin, NDArithmeticMixin, NDSlicingMixin, NDData
 ):
-    """Allows "windowed" access to some properties of an ``NDAstroData``
+    """Provide "windowed" access to some properties of an ``NDAstroData``
     instance.  In particular, ``data``, ``uncertainty``, ``variance``, and
     ``mask`` return clipped data.
     """
@@ -281,17 +301,14 @@ class NDWindowingAstroData(
 
     @property
     def wcs(self):
-        # pylint: disable=protected-access
         return self._target._slice_wcs(self._window)
 
     @property
     def data(self):
-        # pylint: disable=protected-access
         return self._target._get_simple("_data", section=self._window)
 
     @property
     def uncertainty(self):
-        # pylint: disable=protected-access
         return self._target._get_uncertainty(section=self._window)
 
     @property
[…]
