### 2024-08-15 e993718
docs: ✍️ Update and unify docstrings for utils.
