+++ astrodata/utils.py
@@ -32,7 +32,7 @@ warnings.simplefilter("always", AstroDataDeprecationWarning)
 
 
 def deprecated(reason):
-    """Marks a function as deprecated.
+    """Mark a function as deprecated.
 
     Parameters
     ----------
@@ -50,6 +50,7 @@ def deprecated(reason):
     >>> @deprecated("Use another function instead")
     ... def my_function():
     ...     pass
+
     """
 
     def decorator_wrapper(fn):
@@ -97,7 +98,9 @@ def normalize_indices(slc, nitems):
 
 
 class TagSet(namedtuple("TagSet", "add remove blocked_by blocks if_present")):
-    """Named tuple that is used by tag methods to return which actions should
+    """A named tuple of sets of tag strings.
+
+    Named tuple that is used by tag methods to return which actions should
     be performed on a tag set.
 
     All the attributes are optional, and any combination of them can be used,
@@ -128,7 +131,7 @@ class TagSet(namedtuple("TagSet", "add remove blocked_by blocks if_present")):
         This TagSet will be applied only *all* of these tags are present
 
     Examples
-    ---------
+    --------
     >>> TagSet()  # doctest: +SKIP
     TagSet(
         add=set(),
@@ -153,6 +156,15 @@ class TagSet(namedtuple("TagSet", "add remove blocked_by blocks if_present")):
         blocks=set(),
         if_present=set()
     )
+
+
+    Notes
+    -----
+    If arguments are not provided, the default is an empty set.
+
+    These arguments are not applied within the object, instead they are
+    used when tags are being applied to an AstroData object.
+
     """
 
     def __new__(
@@ -163,6 +175,7 @@ class TagSet(namedtuple("TagSet", "add remove blocked_by blocks if_present")):
         blocks=None,
         if_present=None,
     ):
+        """Instantiate a new TagSet object."""
         return super().__new__(
             cls,
             add or set(),
@@ -182,7 +195,7 @@ def astro_data_descriptor(fn):
         The method to be decorated
 
     Returns
-    --------
+    -------
     The tagged method (not a wrapper)
 
     Warning
@@ -238,7 +251,7 @@ def returns_list(fn):
         The method to be decorated
 
     Returns
-    --------
+    -------
     Callable
         A function
 
@@ -319,7 +332,7 @@ def astro_data_tag(fn):
         The method to be decorated
 
     Returns
-    --------
+    -------
     A wrapper function
 
     Notes
@@ -371,9 +384,10 @@ def astro_data_tag(fn):
 
 
 class Section(tuple):
-    """A class to handle n-dimensional sections"""
+    """A class to handle n-dimensional sections."""
 
     def __new__(cls, *args, **kwargs):
+        """Instantiate a new Section object."""
         # Ensure that the order of keys is what we want
         axis_names = [x for axis in "xyzuvw" for x in (f"{axis}1", f"{axis}2")]
 
@@ -397,18 +411,22 @@ class Section(tuple):
 
     @property
     def axis_dict(self):
+        """Return a dictionary with the axis names as keys."""
         return dict(zip(self._axis_names, self))
 
     def __getnewargs__(self):
+        """Return the arguments needed to create a new instance of the object."""
         return tuple(self)
 
     def __getattr__(self, attr):
+        """Check for attrs in the axis_dict."""
         if attr in self._axis_names:
             return self.axis_dict[attr]
 
         raise AttributeError(f"No such attribute '{attr}'")
 
     def __repr__(self):
+        """Return a string representation of the Section object."""
         return (
             "Section("
             + ", ".join([f"{k}={self.axis_dict[k]}" for k in self._axis_names])
@@ -427,7 +445,7 @@ class Section(tuple):
 
     @staticmethod
     def from_string(value):
-        """The inverse of __str__, produce a Section object from a string."""
[…]
