### 2024-08-15 b710437
docs: ✍️ Add examples and clarification to Section
