+++ astrodata/core.py
@@ -1,7 +1,4 @@
-"""This is the core module of the AstroData package. It provides the
-`AstroData` class, which is the main interface to manipulate astronomical
-data sets.
-"""
+"""Core module of the AstroData package, containing the |AstroData| class."""
 
 import inspect
 import logging
@@ -445,7 +442,7 @@ class AstroData:
         attribute's value) will be interpreted as a descriptor.
 
         Returns
-        --------
+        -------
         tuple of str
         """
         members = inspect.getmembers(
@@ -599,7 +596,7 @@ class AstroData:
         The objects are instances of AstroPy's `astropy.nddata.NDUncertainty`,
         or `None` where no information is available.
 
-        See also
+        See Also
         --------
         variance : The actual array supporting the uncertainty object.
 
@@ -646,8 +643,8 @@ class AstroData:
         For objects that miss uncertainty information, `None` will be provided
         instead.
 
-        See also
-        ---------
+        See Also
+        --------
         uncertainty : The uncertainty objects used under the hood.
 
         Notes
@@ -687,6 +684,14 @@ class AstroData:
         self.nddata.wcs = value
 
     def __iter__(self):
+        """Iterate over the extensions..
+
+        This generator yields the `AstroData` object for each extension.
+
+        Notes
+        -----
+        This will yield the object, once, if it's a single slice.
+        """
         if self.is_single:
             yield self
         else:
@@ -705,7 +710,7 @@ class AstroData:
             An integer or an instance of a Python standard `slice` object
 
         Raises
-        -------
+        ------
         TypeError
             If trying to slice an object when it doesn't make sense (e.g.
             slicing a single slice)
@@ -751,7 +756,7 @@ class AstroData:
             to remove.
 
         Raises
-        -------
+        ------
         IndexError
             If `idx` is out of range.
         """
@@ -773,7 +778,7 @@ class AstroData:
             The attribute's name.
 
         Raises
-        -------
+        ------
         AttributeError
             If the attribute could not be found/computed.
 
@@ -887,7 +892,7 @@ class AstroData:
             An attribute name.
 
         Returns
-        --------
+        -------
         bool
         """
         return attribute in self.exposed
@@ -912,7 +917,7 @@ class AstroData:
         added dynamically).
 
         Examples
-        ---------
+        --------
         >>> ad[0].exposed  # doctest: +SKIP
         set(['OBJMASK', 'OBJCAT'])
 
@@ -930,7 +935,7 @@ class AstroData:
         about a single extension, until all extensions have been yielded.
 
         Yields
-        -------
+        ------
         dict
             A dictionary with the pixel information for an extension.
         """
@@ -1096,46 +1101,46 @@ class AstroData:
         )
 
     @format_doc(_ARIT_DOC, name="addition", op="+")
-    def __add__(self, oper):
+    def __add__(self, oper):  # noqa
         copy = deepcopy(self)
         copy += oper
         return copy
 
     @format_doc(_ARIT_DOC, name="subtraction", op="-")
-    def __sub__(self, oper):
+    def __sub__(self, oper):  # noqa
         copy = deepcopy(self)
         copy -= oper
         return copy
 
     @format_doc(_ARIT_DOC, name="multiplication", op="*")
-    def __mul__(self, oper):
+    def __mul__(self, oper):  # noqa
         copy = deepcopy(self)
         copy *= oper
         return copy
 
     @format_doc(_ARIT_DOC, name="division", op="/")
-    def __truediv__(self, oper):
+    def __truediv__(self, oper):  # noqa
         copy = deepcopy(self)
         copy /= oper
         return copy
 
[…]
