### 2024-08-15 b5dd4b5
docs: ✍️ Fix linting errors for fits.py
