+++ astrodata/fits.py
@@ -67,23 +67,40 @@ class FitsHeaderCollection:
         self._headers = list(headers)
 
     def _insert(self, idx, header):
+        """Insert a header at a given index.
+
+        Arguments
+        ---------
+        idx : int
+            The index where to insert the header.
+
+        header : `astropy.io.fits.Header`
+            The header to insert.
+
+        Returns
+        -------
+        None
+        """
         self._headers.insert(idx, header)
 
     def __iter__(self):
+        """Iterate over the headers."""
         yield from self._headers
 
     def __setitem__(self, key, value):
+        """Set keyword value in all the headers."""
         if isinstance(value, tuple):
             self.set(key, value=value[0], comment=value[1])
         else:
             self.set(key, value=value)
 
     def set(self, key, value=None, comment=None):
-        """Set a keyword in all the headers."""
+        """Set a keyword value in all the headers."""
         for header in self._headers:
             header.set(key, value=value, comment=comment)
 
     def __getitem__(self, key):
+        """Get item from all the headers by key."""
         missing_at = []
         ret = []
         for n, header in enumerate(self._headers):
@@ -121,6 +138,7 @@ class FitsHeaderCollection:
             return vals
 
     def __delitem__(self, key):
+        """Remove key from all the headers."""
         self.remove(key)
 
     def remove(self, key):
@@ -155,6 +173,7 @@ class FitsHeaderCollection:
                 raise KeyError(f"{err.args[0]} at header {n}") from err
 
     def __contains__(self, key):
+        """Return True if key is present in any of the headers."""
         return any(tuple(key in h for h in self._headers))
 
 
@@ -340,8 +359,7 @@ def _process_table(table, name=None, header=None):
 
 
 def card_filter(cards, include=None, exclude=None):
-    """Filter a list of cards, lazily returning only those that match the
-    criteria.
+    """Filter a list of cards, returning only those that match the criteria.
 
     Parameters
     ----------
@@ -358,6 +376,7 @@ def card_filter(cards, include=None, exclude=None):
     ------
     card : tuple
         A card that matches the criteria.
+
     """
     for card in cards:
         if include is not None and card[0] not in include:
@@ -370,16 +389,16 @@ def card_filter(cards, include=None, exclude=None):
 
 
 def update_header(headera, headerb):
-    """Update headera with the cards from headerb, but only if they are
-    different.
+    """Update headera with the cards from headerb if they differ.
 
-    Parameters
-    ----------
+    Arguments
+    ---------
     headera : `astropy.io.fits.Header`
         The header to update.
 
     headerb : `astropy.io.fits.Header`
         The header to update from.
+
     """
     cardsa = tuple(tuple(cr) for cr in headera.cards)
     cardsb = tuple(tuple(cr) for cr in headerb.cards)
@@ -406,7 +425,7 @@ def update_header(headera, headerb):
 
 
 def fits_ext_comp_key(ext):
-    """Returns a pair (int, str) that will be used to sort extensions."""
+    """Return a pair (int, str) that can be used to sort extensions."""
     if isinstance(ext, PrimaryHDU):
         # This will guarantee that the primary HDU goes first
         ret = (-1, "")
@@ -442,7 +461,7 @@ class FitsLazyLoadable:
     """Class to delay loading of data from a FITS file."""
 
     def __init__(self, obj):
-        """Initializes the object.
+        """Initialize the object.
 
         Parameters
         ----------
@@ -472,6 +491,7 @@ class FitsLazyLoadable:
[…]
