+++ astrodata/testing.py
@@ -1,5 +1,5 @@
 # pragma: no cover
-"""Fixtures to be used in tests in DRAGONS"""
+"""Define utility functions and classes for testing purposes."""
 
 import enum
 import functools
@@ -38,13 +38,33 @@ log = logging.getLogger(__name__)
 
 
 class DownloadResult(enum.Enum):
+    """Result status for the download_from_archive function.
+
+    Enum class to store the possible states of the download_from_archive
+    function. The states are used to determine if the function was successful
+    in downloading a file from the archive.
+
+    Attributes
+    ----------
+    SUCCESS : int
+        The download was successful.
+
+    NOT_FOUND : int
+        The file was not found in the archive.
+
+    NONE : int
+        The state is not set.
+    """
+
     SUCCESS = 2
     NOT_FOUND = 1
     NONE = 0
 
 
 class DownloadState:
-    """Singleton class to hold the state of the download_from_archive function.
+    """Stores the success state of the download_from_archive function.
+
+    Singleton class to hold the state of the download_from_archive function.
     A bit of an annoying workaround because of conflicts with how ``pytest``'s
     fixtures work.
 
@@ -52,6 +72,27 @@ class DownloadState:
     instantiated directly. Instead, the instance should be accessed via the
     ``_instance`` class attribute. Instantiation using ``DownloadState()`` will
     do this automatically.
+
+    Attributes
+    ----------
+    _state : DownloadResult
+        The state of the download_from_archive function.
+
+    _valid_state : bool
+        Flag to indicate if the state is valid.
+
+    test_result : bool
+        Result of the test download_from_archive function.
+
+    Notes
+    -----
+    To check the state of the download_from_archive function, use the
+    ``check_state`` method. This method will return the state of the function
+    as a :class:`~DownloadResult` enum. If the state is not valid, the method
+    will re-test the function.
+
+    See the :class:`~DownloadResult` enum for the possible states and their
+    documentation.
     """
 
     __slots__ = ["_state", "_valid_state", "test_result"]
@@ -59,6 +100,7 @@ class DownloadState:
     _instance = None
 
     def __new__(cls):
+        """Instantiate or return the singleton class."""
         if cls._instance is None:
             cls._instance = super().__new__(cls)
             cls._instance._state = None
@@ -97,8 +139,7 @@ class DownloadState:
 
 
 def skip_if_download_none(func):
-    """Skip test if download_from_archive is returning None. Otherwise,
-    continue.
+    """Skip test if download_from_archive is returning None.
 
     Used as a wrapper for testing functions. Works with nose, pynose, and
     pytest.
@@ -119,7 +160,9 @@ def skip_if_download_none(func):
 
 
 def get_corners(shape):
-    """This is a recursive function to calculate the corner indices
+    """Calculate the corner indices of an array of the specified shape.
+
+    This is a recursive function to calculate the corner indices
     of an array of the specified shape.
 
     Parameters
@@ -156,9 +199,17 @@ def get_corners(shape):
 
 
 def assert_most_close(
-    actual, desired, max_miss, rtol=1e-7, atol=0, equal_nan=True, verbose=True
+    actual,
+    desired,
+    max_miss,
+    rtol=1e-7,
+    atol=0,
+    equal_nan=True,
+    verbose=True,
 ):
-    """Raises an AssertionError if the number of elements in two objects that
+    """Assert that two objects are equal up to a specified number of elements.
+
+    Raises an AssertionError if the number of elements in two objects that
     are not equal up to desired tolerance is greater than expected.
 
     See Also
@@ -227,11 +278,13 @@ def assert_most_close(
 
 
[…]
