### 2024-08-15 d968cad
docs: ✍️ Add information to package docstring.
