### 2024-08-15 5a8ad41
docs(quickstart): 🚧 Tweak example classes.
