### 2024-08-16 GeminiDRSoftware/astrodata#16 @teald
This is frequent enough---and will be for a bit I suspect---that it's probably better to have some of the failing downloads `xfail` and have that escalate in special cases (e.g., it must pass before a release, or a PR). Usually one/two re-tries will succeed without issue.

If there's a straightforward way to have the action retry itself, even better.
