### 2024-08-16 750cfac
Update API docstrings (#50)
This commit squashed together 26 commits:
* refactor: :wastebasket: Deprecate "add_header_to_table"

* docs: ✍️ State deprecation of add_header_to_table in summary

* docs: ✍️ Update docstring for descriptor decorator.

* docs: ✍️ Update astro_data_tag documentation.

* docs: ✍️ Update returns_list documentation.

* docs(api): ✍️ Update from_file documentation

* docs(api): ✍️ Update open docstring

* docs: 👷 Add nox, install help.

* docs: ✍️ Unify and expand on docstrings in AstoData

* docs: ✍️ Unify and update nddata.py docs

* docs: ✍️ Update and unify docstrin
[…]
