+++ astrodata/wcs.py
@@ -36,16 +36,17 @@ re_cd = re.compile(r"^CD(\d+)_\d+$", re.IGNORECASE)
 # FITS-WCS -> gWCS
 # -----------------------------------------------------------------------------
 def pixel_frame(naxes, name="pixels"):
-    """Make a CoordinateFrame for pixels
+    """Make a CoordinateFrame for pixels.
 
-    Parameters
-    ----------
+    Arguments
+    ---------
     naxes: int
         Number of axes
 
     Returns
     -------
     CoordinateFrame
+        The pixel frame.
     """
     axes_names = ("x", "y", "z", "u", "v", "w")[:naxes]
     return cf.CoordinateFrame(
@@ -59,11 +60,13 @@ def pixel_frame(naxes, name="pixels"):
 
 
 def fitswcs_to_gwcs(input_data, *, raise_errors: bool = False):
-    """Create and return a gWCS object from a FITS header or NDData object.  If
+    """Convert a FITS WCS header or NDData object to a gWCS object.
+
+    Create and return a gWCS object from a FITS header or NDData object.  If
     it can't construct one, it should quietly return None.
 
-    Parameters
-    ----------
+    Arguments
+    ---------
     input_data : `astropy.io.fits.Header` or `astropy.nddata.NDData`
         FITS Header or NDData object with basic FITS WCS keywords.
 
@@ -194,13 +197,15 @@ def fitswcs_to_gwcs(input_data, *, raise_errors: bool = False):
 # TODO: Rename this and deprecate this function. The name implies it is a
 # gwcs object being passed, but it requires an NDData object.
 def gwcs_to_fits(ndd, hdr=None):
-    """Convert a gWCS object to a collection of FITS WCS keyword/value pairs,
+    """Convert a gWCS object to FITS WCS keyword/value pairs.
+
+    Convert a gWCS object to a collection of FITS WCS keyword/value pairs,
     if possible. If the FITS WCS is only approximate, this should be indicated
     with a dict entry {'FITS-WCS': 'APPROXIMATE'}. If there is no suitable FITS
     representation, then a ValueError or NotImplementedError can be raised.
 
-    Parameters
-    ----------
+    Arguments
+    ---------
     ndd : `astropy.nddata.NDData`
         The NDData whose wcs attribute we want converted
 
@@ -211,6 +216,11 @@ def gwcs_to_fits(ndd, hdr=None):
     -------
     dict
         values to insert into the FITS header to express this WCS
+
+    Warning
+    -------
+    This function is experimental and may not work for all WCS objects.
+    It is likely to be deprecated in a future release.
     """
     if hdr is None:
         hdr = {}
@@ -502,7 +512,9 @@ def gwcs_to_fits(ndd, hdr=None):
 
 
 def model_is_affine(model):
-    """Test a Model for affinity. This is currently done by checking the name
+    """Return True if the model is a valid affine transformation.
+
+    Test a Model for affinity. This is currently done by checking the name
     of its class (or the class names of all its submodels)
 
     TODO: Is this the right thing to do? We could compute the affine matrices
@@ -530,15 +542,17 @@ def model_is_affine(model):
 
 
 def calculate_affine_matrices(func, shape, origin=None):
-    """Compute the matrix and offset necessary of an affine transform that
+    """Calculate the affine matrices and offset for a function.
+
+    Compute the matrix and offset necessary of an affine transform that
     represents the supplied function. This is done by computing the linear
     matrix along all axes extending from the centre of the region, and then
     calculating the offset such that the transformation is accurate at the
     centre of the region. The matrix and offset are returned in the standard
     python order (i.e., y-first for 2D).
 
-    Parameters
-    ----------
+    Arguments
+    ---------
     func : callable
         function that maps input->output coordinates
 
[…]
