+++ docs/quickstart.rst
@@ -353,6 +353,8 @@ class as ``'GMOS'`` and ``'SCIENCE'``:
     # Let's remove the GMOSScienceAstroData class from the factory to avoid conflicts.
     factory.remove_class(GMOSScienceAstroData)
 
+    from astrodata import astro_data_tag, TagSet
+
     class GMOSAstroDataTagged(GMOSAstroData):
         """A class for GMOS science data with tags.
 
@@ -374,18 +376,6 @@ class as ``'GMOS'`` and ``'SCIENCE'``:
         def _tag_arc(self):
             if self.phu.get("OBSTYPE") == "ARC":
                 return TagSet(["ARC", "CAL"])
-
-        def _tag_is_bias(self):
-            if self.phu.get("OBSTYPE") == "BIAS":
-                return True
-            else:
-                return False
-
-        def _tag_is_bpm(self):
-            if self.phu.get("OBSTYPE") == "BPM" or "BPMASK" in self.phu:
-                return True
-            else:
-                return False
 
         @astro_data_tag
         def _tag_bias(self):
@@ -422,6 +412,30 @@ class as ``'GMOS'`` and ``'SCIENCE'``:
             else:
                 return TagSet(['SPECT'])
 
+        def _tag_is_bias(self):
+            if self.phu.get("OBSTYPE") == "BIAS":
+                return True
+            else:
+                return False
+
+        def _tag_is_bpm(self):
+            if self.phu.get("OBSTYPE") == "BPM" or "BPMASK" in self.phu:
+                return True
+            else:
+                return False
+
+        def _tag_is_spect(self):
+            pairs = (
+                ('MASKTYP', 0),
+                ('MASKNAME', 'None'),
+                ('GRATING', 'MIRROR')
+            )
+
+            matches = (self.phu.get(kw) == value for (kw, value) in pairs)
+            if any(matches):
+                return False
+            return True
+
     factory.add_class(GMOSAstroDataTagged)
 
 
@@ -436,11 +450,76 @@ Let's make new |AstroData| objects for our files and see what tags they have:
 
 .. code-block:: python
 
+    all_ad_data = []
+
     for f in files:
-        ad = astrodata.from_file(f'quickstart_data/{f}')
+        location = f"quickstart_data/{f}"
+        ad = astrodata.from_file(location)
         print(f"Opened {ad.filename} with class {ad.__class__}")
         print(f"Tags: {ad.tags}")
 
+        all_ad_data.append(ad)
+
+The result:
+
+.. code-block:: text
+
+    Opened N20170614S0201.fits with class <class '__main__.GMOSAstroDataTagged'>
+    Tags: {'IMAGE', 'GMOS'}
+    Opened N20170614S0202.fits with class <class '__main__.GMOSAstroDataTagged'>
+    Tags: {'IMAGE', 'GMOS'}
+    Opened N20170614S0203.fits with class <class '__main__.GMOSAstroDataTagged'>
+    Tags: {'IMAGE', 'GMOS'}
+    Opened N20170614S0204.fits with class <class '__main__.GMOSAstroDataTagged'>
+    Tags: {'IMAGE', 'GMOS'}
+    Opened N20170614S0205.fits with class <class '__main__.GMOSAstroDataTagged'>
+    Tags: {'IMAGE', 'GMOS'}
[…]
