### 2024-08-19 1d03827
docs(quickstart): ✍️ Add table with tag args
