### 2024-08-19 271cf25
Add tags example to quickstart (#51)
This commit squashed 6 commits:
* docs(quickstart): 🚧 Add first draft of tags exmaple.

* docs(quickstart): 🚧 Tweak example classes.

* docs(quickstart): ✍️ First full draft of quickstart.

* docs(quickstart): ✍️ Add table with tag args

* docs(nox): ✨ Add link to end of docs session.

* docs: ✍️ Add clarifying comments to examples.
