### 2024-08-19 6ce0b98
docs: ❌ Remove sphinx_rtd_dark_mode
