+++ pyproject.toml
@@ -30,7 +30,6 @@ tomlkit = "^0.13.0"
 sphinx = "^7.2.6"
 sphinx-automodapi = "^0.17.0"
 sphinx-autoapi = "^3.0.0"
-sphinx-rtd-dark-mode = "^1.3.0"
 sphinx-rtd-theme = "^2.0.0"
 myst-parser = "^2.0.0"
 tomli = "^2.0.1"
