### 2024-08-19 75ccb18
docs: ❌ Remove sphinx_rtd_dark_mode
