+++ docs/conf.py
@@ -36,7 +36,6 @@ extensions = [
     # "sphinx.ext.autosummary",
     "sphinx_automodapi.automodapi",
     "sphinx_automodapi.smart_resolver",
-    "sphinx_rtd_dark_mode",
 ]
 
 
