### 2024-08-19 6a6ae6b
chore: Add URLs to pyproject.toml
