### 2024-08-19 8f74d89
chore: Add URLs to pyproject.toml
