### 2024-08-22 ca09b17
docs(misc): 🚧 Flesh out initial draft.
