+++ docs/miscellaneous/pyopensci_response_1.md
@@ -43,20 +43,21 @@ These are in order of reading the responses in the checklist above, with the ~se
 
 ## Issues
 All issues raised as part of the initial review have been addressed:
+
 <details>
 
 <summary>Issues raised as part of this review</summary>
 
-+ Issue #18
-+ Issue #19
-+ Issue #22
-+ Issue #23
-+ Issue #25
-+ Issue #26
-+ Issue #27
-+ Issue #28
-+ Issue #29
-+ Issue #33
++ [Issue #18](https://github.com/GeminiDRSoftware/astrodata/issues/18)
++ [Issue #19](https://github.com/GeminiDRSoftware/astrodata/issues/19)
++ [Issue #22](https://github.com/GeminiDRSoftware/astrodata/issues/22)
++ [Issue #23](https://github.com/GeminiDRSoftware/astrodata/issues/23)
++ [Issue #25](https://github.com/GeminiDRSoftware/astrodata/issues/25)
++ [Issue #26](https://github.com/GeminiDRSoftware/astrodata/issues/26)
++ [Issue #27](https://github.com/GeminiDRSoftware/astrodata/issues/27)
++ [Issue #28](https://github.com/GeminiDRSoftware/astrodata/issues/28)
++ [Issue #29](https://github.com/GeminiDRSoftware/astrodata/issues/29)
++ [Issue #33](https://github.com/GeminiDRSoftware/astrodata/issues/33)
 
 </details>
 
@@ -66,14 +67,33 @@ All PRs have been merged into the main code:
 
 <details>
 
-<summary>List of Pull Requests made as part of this review</summary>
+<summary>List of Pull Requests made by reviewers</summary>
 
-+ Pull Request #20
-+ Pull Request #21
++ [Pull Request #20](https://github.com/GeminiDRSoftware/astrodata/pull/20)
++ [Pull Request #21](https://github.com/GeminiDRSoftware/astrodata/pull/21)
 
 </details>
 
 Thank you for your contributions!
+
+<details>
+
+<summary>List of Pull Requests related to this review.</summary>
+
++ [Pull Request #24](https://github.com/GeminiDRSoftware/astrodata/pull/24)
++ [Pull Request #31](https://github.com/GeminiDRSoftware/astrodata/pull/31)
++ [Pull Request #32](https://github.com/GeminiDRSoftware/astrodata/pull/32)
++ [Pull Request #34](https://github.com/GeminiDRSoftware/astrodata/pull/34)
++ [Pull Request #38](https://github.com/GeminiDRSoftware/astrodata/pull/38)
++ [Pull Request #41](https://github.com/GeminiDRSoftware/astrodata/pull/41)
++ [Pull Request #46](https://github.com/GeminiDRSoftware/astrodata/pull/46)
++ [Pull Request #48](https://github.com/GeminiDRSoftware/astrodata/pull/48)
++ [Pull Request #50](https://github.com/GeminiDRSoftware/astrodata/pull/50)
++ [Pull Request #51](https://github.com/GeminiDRSoftware/astrodata/pull/51)
++ [Pull Request #53](https://github.com/GeminiDRSoftware/astrodata/pull/53)
+
+</details>
+
 
 ## Other changes of note
 
@@ -81,11 +101,8 @@ Thank you for your contributions!
 [poetry devs]: https://geminidrsoftware.github.io/astrodata/developer/index.html#install-the-dependencies
 [test docs]: https://geminidrsoftware.github.io/astrodata/developer/index.html#run-the-tests
 
-1. We've moved from using `tox` for testing to `nox` for testing and general
-automation. This was already planned for after this review to better support
-`conda` testing, but it made more sense to get this done now.
-2. The developer documentation has been overhauled to describe several new
-development features motivated by these reviews:
[…]
