+++ docs/miscellaneous/pyopensci_response_1.md
@@ -17,20 +17,19 @@ These are in order of reading the responses in the checklist above, with the ~se
 <!-- TODO: <details> & <summary> for this section -->
 
 1. (Documentation) Examples missing for user-facing functions
-    1. All documentation in the ["Common API for Users"][api_short] now
-    have examples, better explanations, or are deprecated.
+    1. All documentation in the ["Common API for Users"][api_short] now have examples, better explanations, or are deprecated.
     2. Specifically, [`add_header_to_table`][add_header_to_table] has been deprecated, as its functionality is no longer used. It will be removed in version 3.0.0.
-    3. Covered in PR #50
+    3. Covered in [PR #50](https://github.com/GeminiDRSoftware/astrodata/pull/50)
     4. The previous examples directory was a vestigial directory, and has been removed.
 2. (Documentation) URLs have been added to `pyproject.toml`
-    1. Commit 8f74d89
+    1. Commit [8f74d89](https://github.com/GeminiDRSoftware/astrodata/commit/8f74d89)
 3. (Documentation/README) Badges in README now have a repostatus and Python version badge
-    1. Commit 6537a51
+    1. Commit [6537a51](https://github.com/GeminiDRSoftware/astrodata/commit/6537a51)
 4. (Documentation/README) Citation information has been added to the README
-    1. Commit b057511
+    1. Commit [b057511](https://github.com/GeminiDRSoftware/astrodata/commit/b057511)
 5. (Functionality/Automated tests) Testing failures
     1. All testing failures encountered were due to intermittent service with the Gemini Archive (which is getting an upgrade!), as well as some issues with Actions and their runner configurations.
-    2. Tracked in Issue #16 until updates complete
+    2. Tracked in [Issue #16](https://github.com/GeminiDRSoftware/astrodata/commit/b057511) until updates complete
 6. (Functionality/Packaging Instructions) Testing instructions unclear
     1. [Developer documentation][testing docs] has been upgraded to reflect our new testing framework (using `nox`).
     2. Includes instructions for selecting specific tests, which tests are default, and specify python versions tested.
