### 2024-08-23 75f07ba
ci(pre-commit): 🛠️ Update default stages
