### 2024-08-23 b99d1a8
ci: 🛠️ Allow build & release to trigger and wait on tests.
