+++ .github/workflows/source_tests.yml
@@ -8,6 +8,7 @@ on:
     schedule:
         - cron: "20 1 * * sat"
     workflow_dispatch:
+    workflow_call:
 
 defaults:
     run:
