### 2024-08-23 2101946
ci: 🚀 run on push to allow workflow triggering.
