### 2024-08-23 4922305
ci: ♻️ restore push behavior w/tags
