### 2024-08-23 fbcb124
ci: 🐛 Fix if statemetns for workflow_dispatch
