### 2024-08-23 440e0a5
ci: 🚀 Uncomment publish statement for non-workflow_dispatch runs
