### 2024-08-23 cee5fe2
ci: 🛠️ Disallow minting tokens on workflow dispatch.
