### 2024-08-26 91ab72e
ci: :clock: Schedule for Mondays
