### 2024-08-26 b891fd2
[FIX] Have publishing workflow test before publishing (#55)
Squash of 6 commits:
* ci: 🛠️ Allow build & release to trigger and wait on tests.

* ci: 🚀 run on push to allow workflow triggering.

* ci: ♻️ restore push behavior w/tags

* ci: 🐛 Fix if statemetns for workflow_dispatch

* ci: 🚀 Uncomment publish statement for non-workflow_dispatch runs

* ci: 🛠️ Disallow minting tokens on workflow dispatch.

* chore(README): 🎨 Space in README
