+++ .github/workflows/build_tests.yml
@@ -8,6 +8,7 @@ on:
     schedule:
         - cron: "20 1 * * mon"
     workflow_dispatch:
+    workflow_call:
 
 defaults:
     run:
