### 2024-08-26 c2fbe6b
docs(misc): 🚧 Add clarification to gwcs comment.
