### 2024-08-27 GeminiDRSoftware/astrodata#57 @teald
[DOC] Create a more complete user example showcasing non-developer utilities
# Summary

[pyopensci review]: https://github.com/pyOpenSci/software-submission/issues/181#issuecomment-2229579602

The [PyOpenSci review][pyopenscireview] brought up that the examples in `astrodata` primarily focus on adding, manipulating, and utilizing tags and descriptors. Since `astrodata` has other features that are useful in data processing workflows, it'd be good to have an example that showcases working with data as a whole, using only `astrodata` & its dependencies.

# To-do

- [ ]: Outline example process(es) and features they highlight
- [ ]: Source data (fake/generated or real, ideally non-Gemini)
- [ ]: Check what might be consolidated from User/Programmer Manuals (see Issue #47)
