### 2024-08-27 246606d
docs(README): ✨ Add feature list, links to manuals to README
