### 2024-08-27 21103a9
docs(README): ✨ Add feature list, links to manuals to README
