+++ astrodata/wcs.py
@@ -306,12 +306,32 @@ def gwcs_to_fits(ndd, hdr=None):
         # Remove projection parts so we can calculate the CD matrix
         if projcode:
             nat2cel.name = "nat2cel"
+            transform_inverse = transform.inverse.copy()
+
+            for m in transform_inverse:
+                if isinstance(m, models.RotateCelestial2Native):
+                    m.name = "cel2nat"
+
+                elif isinstance(m, models.Sky2PixProjection):
+                    m.name = "sky2pix"
+
+            # TODO(teald): Below comments are from DRAGONS commit 46177cc
+            # transform_inverse = transform_inverse.replace_submodel(
+            #     "cel2nat", models.Identity(2)
+            # )
+            # transform_inverse = transform_inverse.replace_submodel(
+            #     "sky2pix", models.Identity(2)
+            # )
+
             transform = transform.replace_submodel(
                 "pix2sky", models.Identity(2)
             )
             transform = transform.replace_submodel(
                 "nat2cel", models.Identity(2)
             )
+
+            # TODO(teald): Below comments are from DRAGONs commit 46177cc
+            # transform.inverse = transform_inverse
 
     # Replace a log-linear axis with a linear axis representing the log
     # and a Tabular axis with Identity to ensure the affinity check is passed
@@ -421,12 +441,13 @@ def gwcs_to_fits(ndd, hdr=None):
 
     crval = [wcs_dict[f"CRVAL{i+1}"] for i, _ in enumerate(world_axes)]
 
-    try:
-        crval[lon_axis] = 0
-        crval[lat_axis] = 0
+    # TODO: Commented out in DRAGONS commit 46177cc
+    # try:
+    #     crval[lon_axis] = 0
+    #     crval[lat_axis] = 0
 
-    except NameError:
-        pass
+    # except NameError:
+    #     pass
 
     # Find any world axes that we previous logarithmed and fix the CDij
     # matrix -- we follow FITS-III (Greisen et al. 2006; A&A 446, 747)
@@ -445,9 +466,10 @@ def gwcs_to_fits(ndd, hdr=None):
 
             crval[world_axis - 1] = np.log(crval[world_axis - 1])
 
+    # TODO: Below comment may not apply, from DRAGONS commit 46177cc
     # This (commented) line fails for un-invertable Tabular2D
-    # crpix = np.array(wcs.backward_transform(*crval)) + 1
-    crpix = np.array(transform.inverse(*crval)) + 1
+    crpix = np.array(wcs.backward_transform(*crval)) + 1
+    # crpix = np.array(transform.inverse(*crval)) + 1
 
     # Cope with a situation where the sky projection center is not in the slit
     # We may be able to fix this in future, but FITS doesn't handle it well.
@@ -948,9 +970,23 @@ def fitswcs_image(header):
     # create a "ghost" orthogonal axis here so an inverse can be defined
     # Modify the CD matrix in case we have to use a backup Matrix Model later
     if len(pixel_axes) == 1:
-        cd[sky_axes[0], -1] = -cd[sky_axes[1], pixel_axes[0]]
-        cd[sky_axes[1], -1] = cd[sky_axes[0], pixel_axes[0]]
-        sky_cd = cd[np.ix_(sky_axes, pixel_axes + [-1])]
+        # TODO: Below commented code is from DRAGONS commit 46177cc
+        sky_cd = np.array(
+            [
+                [
+                    cd[sky_axes[0], pixel_axes[0]],
+                    -cd[sky_axes[1], pixel_axes[0]],
+                ],
+                [
+                    cd[sky_axes[1], pixel_axes[0]],
[…]
