+++ astrodata/wcs.py
@@ -306,7 +306,7 @@ def gwcs_to_fits(ndd, hdr=None):
         # Remove projection parts so we can calculate the CD matrix
         if projcode:
             nat2cel.name = "nat2cel"
-            transform_inverse = transform.inverse.copy()
+            transform_inverse = transform.inverse
 
             for m in transform_inverse:
                 if isinstance(m, models.RotateCelestial2Native):
@@ -315,13 +315,12 @@ def gwcs_to_fits(ndd, hdr=None):
                 elif isinstance(m, models.Sky2PixProjection):
                     m.name = "sky2pix"
 
-            # TODO(teald): Below comments are from DRAGONS commit 46177cc
-            # transform_inverse = transform_inverse.replace_submodel(
-            #     "cel2nat", models.Identity(2)
-            # )
-            # transform_inverse = transform_inverse.replace_submodel(
-            #     "sky2pix", models.Identity(2)
-            # )
+            transform_inverse = transform_inverse.replace_submodel(
+                "cel2nat", models.Identity(2)
+            )
+            transform_inverse = transform_inverse.replace_submodel(
+                "sky2pix", models.Identity(2)
+            )
 
             transform = transform.replace_submodel(
                 "pix2sky", models.Identity(2)
@@ -330,8 +329,7 @@ def gwcs_to_fits(ndd, hdr=None):
                 "nat2cel", models.Identity(2)
             )
 
-            # TODO(teald): Below comments are from DRAGONs commit 46177cc
-            # transform.inverse = transform_inverse
+            transform.inverse = transform_inverse
 
     # Replace a log-linear axis with a linear axis representing the log
     # and a Tabular axis with Identity to ensure the affinity check is passed
@@ -441,18 +439,12 @@ def gwcs_to_fits(ndd, hdr=None):
 
     crval = [wcs_dict[f"CRVAL{i+1}"] for i, _ in enumerate(world_axes)]
 
-    # TODO: Commented out in DRAGONS commit 46177cc
-    # try:
-    #     crval[lon_axis] = 0
-    #     crval[lat_axis] = 0
+    try:
+        crval[lon_axis] = 0
+        crval[lat_axis] = 0
 
-    # except NameError:
-    #     pass
-
-    # TODO: Below comment may not apply, from DRAGONS commit 46177cc
-    # This (commented) line fails for un-invertable Tabular2D
-    crpix = np.array(wcs.backward_transform(*crval)) + 1
-    # crpix = np.array(transform.inverse(*crval)) + 1
+    except NameError:
+        pass
 
     # Find any world axes that we previous logarithmed and fix the CDij
     # matrix -- we follow FITS-III (Greisen et al. 2006; A&A 446, 747)
@@ -470,6 +462,11 @@ def gwcs_to_fits(ndd, hdr=None):
                 )
 
             crval[world_axis - 1] = np.log(crval[world_axis - 1])
+
+    # TODO: Below comment may not apply, from DRAGONS commit 46177cc
+    # This (commented) line fails for un-invertable Tabular2D
+    # crpix = np.array(wcs.backward_transform(*crval)) + 1
+    crpix = np.array(transform.inverse(*crval)) + 1
 
     # Cope with a situation where the sky projection center is not in the slit
     # We may be able to fix this in future, but FITS doesn't handle it well.
