### 2024-09-09 e97fab5
docs(config): 🧹 Remove Anaconda reference
This removes a remaining Anaconda reference in docs/conf.py, which means
it should show up in "nitpicky" build mode for the docs if there are any
remaining links.
