### 2024-09-10 17a0a04
chore(dragons update): 📦 Update docstring
Covers DRAGONS commit 11a499e5c89d436ebe030a3ccc07c0ed65946750. This
ignores the regex update because that's already been fixed using
r-strings.
