### 2024-09-27 GeminiDRSoftware/astrodata#16 @teald
Updates to the archive should have this solved 🤞
