### 2024-09-28 #181 @hamogu
I see a very detailed reply from the package author thanks to @teald for explaining so detailed what work you have done to address the review!

@aaryapatil and @mwcraig: Please have a look and if you agree that that addresses all your concerns, please tick the box "The author has responded to my review and made changes to my satisfaction. I recommend approving this package." by editing the message of your review (and I appreciate a small note here, because GH will notify me if you post a new comment, but not if you edit an existing comment); if there is more to be done, please reply here so that we all know.
