### 2024-11-12 #181 @hamogu
@aaryapatil and @mwcraig : I know we all can get busy at times, but it would be really great if you could check if the changes and replies by @teald address your concerns. If you think that the package should be accepted now, please edit your response above to  tick the box "The author has responded to my review and made changes to my satisfaction. I recommend approving this package." by editing the message of your review (and I appreciate a small note here, because GH will notify me if you post a new comment, but not if you edit an existing comment); if there is more to be done, please reply here so that we all know.

If for some reason, you can't get to check that in the next two weeks or so, please let me know, so that we can proceed with this review in some other way! Thanks for your help and apologies to @teald that I'm letting you wait this long!
