### 2024-12-30 #181 @mwcraig
Apologies for this long-delayed re-review. I really appreciate ll of the effort put in by the authors to address the comments in my review.

Looking it over again this morning I only saw a couple remaining issues, both related to testing. Please see https://github.com/pyOpenSci/software-submission/issues/181#issuecomment-2229579602 for details -- look for the red exclamation marks.

These should be easy fixes, I think...
