### 2025-01-31 #181 @hamogu
@teald I absolutely apologize for having this dragged out to much. Usual policy in pyOpenSci is that reviews are down in 4 weeks and any later iteration only takes 2-4 weeks. By this measure, this issue would have been done in July or August. It's absolutely my fault as the editor that I let this drag out so long. I'm sorry and I promise I'll learn from it and organize the review in a different way if reviewers are not available because their live is busy.

That said, I'd really like to get this to a close. Do you think you'll have time to fix the few remaining test failures that are listed in https://github.com/pyOpenSci/software-submission/issues/181#issuecomment-2229579602 ?
