### 2025-01-31 #181 @teald
Thanks @hamogu for helping us out with this!

I apologize for the delay, I was on break when the response came in and we've been dealing with external issues with accessing databases through our CI/CD workflows. I was hoping those issues would be resolved sooner rather than later, but I can offer a preliminary response now at least:

@mwcraig:

> tests/integration/dragons/test_calibration_setup.py fails at import gemini_instruments because there is no module gemini_instruments. I do see that this test is not included in the ones run by nox, so perhaps this one is not expected to work now?

The integration tests require DRAGONS to be installed, and it gets installed during `nox -s dragons_release_tests` (tests using the current DRAGONS release) or `nox -s dragons_dev_tests` (tests using the version of DRAGONS on github). Do those sessions fail when you run them?

This is mostly meant for our CI/CD, not really developers, which is why the unit tests are the focus of the developer documentation. But they should still be able to run locally.

>  tests/unit/test_wcs.py raises a warning at line 269: PytestUnknownMarkWarning: Unknown pytest.mark.preprocessed_data - is this a typo?.... This warning is generated both when running pytest and when running the unit tests via nox -s unit_tests-3.12.

This is fixed in the PR https://github.com/GeminiDRSoftware/astrodata/pull/82. The tests there are failing because of the aforementioned external issue, which is why it's not yet integrated into the main codebase.
