### 2025-02-04 #181 @hamogu
Thanks @teald! As an editor, I'm deciding that that is good enough for acceptance; I don't want to risk another long wait while we are all busy. I know you are continuing to work on this package!

🎉 astrodata has been approved by pyOpenSci! Thank you @teald for submitting astrodata and many thanks to @mwcraig and @aaryapatil  for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the README.md of astrodata. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/181)`.

---
