### 2025-02-04 #181 @hamogu
@teald: As you see in the final checklist, we really, really want the package to be archived in a way that it has a DOI and institutional backing to be freely archived for a long time. Github is owned by Microsoft and they could choose to make access to any repro paywall only any day. Zenodo is the easiest solution for that - see https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

Also, I would like to invite all package authors and reviewers to join the pyOpenSci slack space. There is a lot of discussion on best practices for python packaging and a surprisingly large number of helpful people if you ver get stuck somewhere with packaging. If you are interested, just send my the email address I should be inviting, e.g. by emailing me (I think you all know my email from some interaction) to contacting me in e.g. the astropy slack with a PM.
