### 2025-02-04 #181 @teald
All the above author wrap up tasks are completed:

+ [DOI](https://doi.org/10.5281/zenodo.14804282)
+ [README.md now has the requested badges](https://github.com/GeminiDRSoftware/astrodata/blob/6a59dc695757f231117599e47869b4d3a6c5af9d/README.md)
+ I've filled out the post-review survey

Let me know if there's anything else you need from me!
