### 2025-02-04 #181 @hamogu
Hi, I think we're good. If you want to write a blog post for PyOpenSci highlighting your package [like this example](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis) then you are welcome - but that's totally optional. I don't know what range a PyOpenSci blog post has on google, and I'm sure you are busy. but if you want to write one, get in touch.

Otherwise, thank you so much for your work in getting this package through. It might take a day to appear on the pyOpenSci listing (I think it's a nightly build of the website) and more than that if I messed up the formatting of the json that gets parsed - but I'll check and fix that if it's not appearing tomorrow.
