### 2025-04-11 #181 @lwasser
hi there, Colleagues! It appears as if this review is complete and can be closed 🚀 congratulations on being accepted to both pyOpenSci and affiliated with Astropy!! I am going to go ahead and close it HOWEVER, please do reopen if it should remain open!!
